refactor(coin-bitcoin): move zcash impl to coin-bitcoin
wip
