+++ .changeset/fresh-geese-tie.md
@@ -1,5 +1,4 @@
 ---
-"@ledgerhq/zcash-shielded": minor
 "@ledgerhq/coin-bitcoin": minor
 "ledger-live-desktop": minor
 ---
