+++ .changeset/nasty-snails-fold.md
@@ -1,5 +1,4 @@
 ---
-"@ledgerhq/zcash-shielded": minor
 "@ledgerhq/coin-bitcoin": minor
 ---
