+++ .changeset/silver-frogs-occur.md
@@ -1,5 +1,4 @@
 ---
-"@ledgerhq/zcash-shielded": minor
 "@ledgerhq/coin-bitcoin": minor
 "ledger-live-desktop": minor
 ---
