+++ apps/ledger-live-desktop/package.json
@@ -71,7 +71,6 @@
     "@ledgerhq/coin-cosmos": "workspace:^",
     "@ledgerhq/coin-evm": "workspace:^",
     "@ledgerhq/coin-filecoin": "workspace:^",
-    "@ledgerhq/zcash-shielded": "workspace:^",
     "@ledgerhq/zcash-utils": "catalog:",
     "@ledgerhq/coin-module-framework": "catalog:",
     "@ledgerhq/crypto-icons": "catalog:",
