+++ apps/ledger-live-desktop/src/renderer/families/bitcoin/ZCashExportKeyFlowModal/Body.tsx
@@ -17,7 +17,7 @@ import { BitcoinAccount } from "@ledgerhq/live-common/families/bitcoin/types";
 import { ContinueFooter } from "./shared/ContinueFooter";
 import StepUfvk from "./steps/StepUfvk";
 import StepDevice, { StepDeviceFooter } from "./steps/StepDevice";
-import { ZcashSyncState } from "@ledgerhq/zcash-shielded/types";
+import { ZcashSyncState } from "@ledgerhq/coin-bitcoin/chain-adapters/zcash/types";
 
 export type Data = {
   account: BitcoinAccount;
