+++ libs/coin-modules/.oxlintrc.json
@@ -86,12 +86,6 @@
       }
     },
     {
-      "files": ["zcash-shielded/**/*.ts"],
-      "rules": {
-        "import/no-cycle": "warn"
-      }
-    },
-    {
       "files": [
         "coin-stellar/src/globals.d.ts",
         "**/coin-stellar/src/globals.d.ts",
