+++ libs/coin-modules/coin-bitcoin/.unimportedrc.json
@@ -40,6 +40,16 @@
     "src/supportedFeatures.ts",
     "src/fixtures/common.fixtures.ts",
     "src/buildRbfTransaction.ts",
+    "src/chain-adapters/zcash/ZCashIPC.ts",
+    "src/chain-adapters/zcash/ipc/contract.ts",
+    "src/chain-adapters/zcash/ipc/main-host.ts",
+    "src/chain-adapters/zcash/ipc/one-shot-router.ts",
+    "src/chain-adapters/zcash/ipc/utility-entry.ts",
+    "src/chain-adapters/zcash/native-engine/engine.ts",
+    "src/chain-adapters/zcash/serialization/rehydrate.ts",
+    "src/chain-adapters/zcash/sync-estimator.ts",
+    "src/chain-adapters/zcash/ZCash.ts",
+    "src/chain-adapters/zcash/stubs/ZCash.native.ts",
     "src/editTransaction/getEditTransactionPatch.ts",
     "src/editTransaction/getFormattedFeeFields.ts",
     "src/editTransaction/getMinEditTransactionFees.ts",
@@ -51,5 +61,5 @@
     "src/operation.ts"
   ],
   "ignoreUnresolved": ["@jest/expect-utils", "@jest/get-type", "jest-util"],
-  "ignoreUnused": []
+  "ignoreUnused": ["@ledgerhq/zcash-utils"]
 }
