+++ libs/coin-modules/coin-bitcoin/package.json
@@ -36,6 +36,14 @@
   "exports": {
     "./lib/*": "./lib/*.js",
     "./lib-es/*": "./lib-es/*.js",
+    "./chain-adapters/zcash/ZCash": {
+      "react-native": "./src/chain-adapters/zcash/stubs/ZCash.native.ts",
+      "@ledgerhq/source": "./src/chain-adapters/zcash/ZCash.ts",
+      "types": "./lib/chain-adapters/zcash/ZCash.d.ts",
+      "import": "./lib-es/chain-adapters/zcash/ZCash.js",
+      "require": "./lib/chain-adapters/zcash/ZCash.js",
+      "default": "./lib-es/chain-adapters/zcash/ZCash.js"
+    },
     "./*": {
       "@ledgerhq/source": "./src/*.ts",
       "require": "./lib/*.js",
@@ -56,7 +64,7 @@
     "@ledgerhq/logs": "workspace:^",
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
-    "@ledgerhq/zcash-shielded": "workspace:^",
+    "@ledgerhq/zcash-utils": "catalog:",
     "bchaddrjs": "^0.5.2",
     "bech32": "^1.1.3",
     "bignumber.js": "^9.1.2",
