+++ libs/coin-modules/coin-bitcoin/src/chain-adapters/zcash/ipc/contract.ts
@@ -10,7 +10,7 @@
  * no Buffer subclasses). Numeric amounts travel as strings in
  * {@link ShieldedSyncResultRaw}; BigNumber reconstruction happens client-side.
  *
- * ─── Error reporting conventions ────────────────────────────────────────────
+ * --- Error reporting conventions ----
  *
  * Two flavours of methods live on this contract, with two different ways of
  * signalling failure. Stick to them when adding a new method so the utility
@@ -36,25 +36,25 @@ import type { ShieldedSyncResultRaw } from "../types";
 /** Unique per-subscription id, generated by the renderer client. */
 export type RequestId = string;
 
-// ─── Renderer <-> Main (ipcRenderer.invoke / webContents.send) ───────────────
+// --- Renderer <-> Main (ipcRenderer.invoke / webContents.send) ---
 
-/** Channel names — centralized to keep renderer/main/utility in lock-step. */
+/** Channel names -- centralized to keep renderer/main/utility in lock-step. */
 export const ZCASH_IPC = {
-  /** invoke → number (chain tip height) */
+  /** invoke -> number (chain tip height) */
   getChainTip: "zcash:getChainTip",
-  /** invoke → number (block height for a given timestamp) */
+  /** invoke -> number (block height for a given timestamp) */
   findBlockHeight: "zcash:findBlockHeight",
-  /** invoke → void once the start-sync message has been posted to the utility. Chunks arrive on `stream`. */
+  /** invoke -> void once the start-sync message has been posted to the utility. Chunks arrive on `stream`. */
   startSync: "zcash:startSync",
-  /** invoke → void. Propagates a cancel signal to the utility. */
+  /** invoke -> void. Propagates a cancel signal to the utility. */
   cancelSync: "zcash:cancelSync",
-  /** webContents.send → StreamEvent. Filter by requestId on the renderer side. */
+  /** webContents.send -> StreamEvent. Filter by requestId on the renderer side. */
   stream: "zcash:stream",
 } as const;
 
 export type ZcashIpcChannel = (typeof ZCASH_IPC)[keyof typeof ZCASH_IPC];
 
-// ─── Request payloads ────────────────────────────────────────────────────────
+// --- Request payloads ---
 
 export type GetChainTipArgs = {
   requestId: RequestId;
@@ -80,7 +80,7 @@ export type CancelSyncArgs = {
   requestId: RequestId;
 };
 
-// ─── Stream events (main → renderer, utility → main) ─────────────────────────
+// --- Stream events (main -> renderer, utility -> main) ---
 
 /**
  * Events emitted during a sync subscription. The renderer client filters them
@@ -92,7 +92,7 @@ export type StreamEvent =
   | { requestId: RequestId; kind: "complete" }
   | { requestId: RequestId; kind: "error"; message: string };
 
-// ─── Utility <-> Main (parentPort.postMessage) ──────────────────────────────
+// --- Utility <-> Main (parentPort.postMessage) ---
 //
 // The utility-facing protocol is symmetric: the main process forwards the same
 // `StartSyncArgs` / `CancelSyncArgs` / `GetChainTipArgs` shapes down the
