--- libs/coin-modules/zcash-shielded/.unimportedrc.json
@@ -1,29 +1,0 @@
-{
-  "entry": [
-    "src/ZCashNative.ts",
-    "src/ZCashNativeIPC.ts",
-    "src/ipc/main-host.ts",
-    "src/ipc/utility-entry.ts"
-  ],
-  "ignorePatterns": [
-    "**/node_modules/**",
-    "**/*.fixture.ts",
-    "**/*.mock.ts",
-    "**/*.test.{js,jsx,ts,tsx}"
-  ],
-  "ignoreUnresolved": [
-    "@jest/expect-utils",
-    "@jest/get-type",
-    "jest-util",
-    "electron"
-  ],
-  "ignoreUnused": [
-    "@ledgerhq/coin-bitcoin",
-    "bignumber.js"
-  ],
-  "ignoreUnimported": [
-    "src/constants.ts",
-    "src/types.ts",
-    "src/stubs/ZCashNative.native.ts"
-  ]
-}
