--- libs/coin-modules/zcash-shielded/README.md
@@ -1,11 +1,0 @@
-<img src="https://user-images.githubusercontent.com/4631227/191834116-59cf590e-25cc-4956-ae5c-812ea464f324.png" height="100" />
-
-[GitHub](https://github.com/LedgerHQ/ledger-live/),
-[Ledger Devs Discord](https://developers.ledger.com/discord-pro),
-[Developer Portal](https://developers.ledger.com/)
-
-## @ledgerhq/zcash-shielded
-
-ZCash shielded account library
-
-***
