--- libs/coin-modules/zcash-shielded/jest.config.js
@@ -1,37 +1,0 @@
-module.exports = {
-  collectCoverageFrom: ["src/**/*.ts"],
-  coveragePathIgnorePatterns: [
-    "test/cli.ts",
-    "/src/stubs/",
-    "/src/ipc/",
-    "/src/types\\.ts",
-    ".*\\.integration\\.test\\.[tj]s",
-    ".*\\.integ\\.test\\.[tj]s",
-  ],
-  coverageReporters: ["json", ["lcov", { file: "lcov.info", projectRoot: "../../../" }], "text"],
-  coverageDirectory: "coverage",
-  testEnvironment: "node",
-  roots: ["<rootDir>/src", "<rootDir>/tests"],
-  testPathIgnorePatterns: [
-    "lib/",
-    "lib-es/",
-    ".*\\.integration\\.test\\.[tj]s",
-    ".*\\.integ\\.test\\.[tj]s",
-  ],
-  reporters: [
-    "default",
-    ...(process.env.CI ? ["github-actions"] : []),
-    ["jest-sonar", { outputName: "sonar-executionTests-report.xml", reportedFilePath: "absolute" }],
-  ],
-  transformIgnorePatterns: ["node_modules/(?!cipherscan)/"],
-  transform: {
-    "^.+\\.(t|j)sx?$": [
-      "@swc/jest",
-      {
-        jsc: {
-          target: "esnext",
-        },
-      },
-    ],
-  },
-};
