--- libs/coin-modules/zcash-shielded/jest.integ.config.js
@@ -1,18 +1,0 @@
-/** @type {import('jest').Config} */
-module.exports = {
-  testEnvironment: "node",
-  testRegex: ".integ.test.ts$",
-  testPathIgnorePatterns: ["lib/", "lib-es/"],
-  testTimeout: 60_000,
-  forceExit: true,
-  transform: {
-    "^.+\\.(t|j)sx?$": [
-      "@swc/jest",
-      {
-        jsc: {
-          target: "esnext",
-        },
-      },
-    ],
-  },
-};
