--- libs/coin-modules/zcash-shielded/package.json
@@ -1,84 +1,0 @@
-{
-  "name": "@ledgerhq/zcash-shielded",
-  "version": "0.9.0",
-  "description": "ZCash library shielded accounts",
-  "keywords": [
-    "ZCash"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/LedgerHQ/ledger-live.git"
-  },
-  "bugs": {
-    "url": "https://github.com/LedgerHQ/ledger-live/issues"
-  },
-  "homepage": "https://github.com/LedgerHQ/ledger-live/tree/develop/libs/coin-modules/zcash-shielded",
-  "publishConfig": {
-    "access": "public"
-  },
-  "types": "lib/types.d.ts",
-  "typesVersions": {
-    "*": {
-      "lib/*": [
-        "lib/*"
-      ],
-      "lib-es/*": [
-        "lib-es/*"
-      ],
-      "*": [
-        "lib/*"
-      ]
-    }
-  },
-  "exports": {
-    "./lib/*": "./lib/*.js",
-    "./lib-es/*": "./lib-es/*.js",
-    "./ZCashNative": {
-      "react-native": "./src/stubs/ZCashNative.native.ts",
-      "@ledgerhq/source": "./src/ZCashNative.ts",
-      "types": "./lib/ZCashNative.d.ts",
-      "require": "./lib/ZCashNative.js",
-      "default": "./lib-es/ZCashNative.js"
-    },
-    "./*": {
-      "@ledgerhq/source": "./src/*.ts",
-      "types": "./lib/*.d.ts",
-      "require": "./lib/*.js",
-      "default": "./lib-es/*.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "license": "Apache-2.0",
-  "scripts": {
-    "clean": "rimraf lib lib-es",
-    "build": "tsc --project tsconfig.build.json && tsc --project tsconfig.build.json -m esnext --moduleResolution bundler --outDir lib-es",
-    "coverage": "jest --coverage --passWithNoTests",
-    "prewatch": "pnpm build",
-    "watch": "tsc --watch --project tsconfig.build.json",
-    "watch:es": "tsc --watch --project tsconfig.build.json -m esnext --moduleResolution bundler --outDir lib-es",
-    "doc": "documentation readme src/** --section=API --pe ts --re ts --re d.ts",
-    "lint": "oxlint src",
-    "lint:fix": "oxfmt src && oxlint src --fix --quiet",
-    "test": "jest",
-    "test-integ": "jest --config=jest.integ.config.js",
-    "typecheck": "tsc --noEmit --customConditions node",
-    "unimported": "unimported",
-    "format": "oxfmt src",
-    "format:check": "oxfmt src --check"
-  },
-  "devDependencies": {
-    "@swc/core": "catalog:",
-    "@swc/jest": "catalog:",
-    "@types/jest": "catalog:",
-    "@types/node": "catalog:",
-    "jest": "catalog:",
-    "oxfmt": "catalog:",
-    "oxlint": "catalog:"
-  },
-  "dependencies": {
-    "@ledgerhq/logs": "workspace:^",
-    "@ledgerhq/zcash-utils": "catalog:",
-    "bignumber.js": "^9.1.2",
-    "rxjs": "catalog:"
-  }
-}
