--- libs/coin-modules/zcash-shielded/src/types.ts
@@ -1,134 +1,0 @@
-import type { BigNumber } from "bignumber.js";
-import type { Observable } from "rxjs";
-
-export type SyncShieldedArgs = {
-  startBlockHeight: number;
-  viewingKey: string;
-  maxBatchSize: number;
-};
-
-/**
- * Public surface shared by the two native-engine clients
- * ({@link ../ZCashNative.ZCashNative} — in-process, and
- * {@link ../ZCashNativeIPC.ZCashNativeIPC} — Electron renderer bridge).
- *
- * Both implementations MUST stay signature-compatible so the renderer bundle
- * can alias one to the other (see `rspack.renderer.ts`). Declaring
- * `implements ZCashNativeClient` on each class lets the compiler enforce that
- * symmetry — if a method is added or renamed on one side without the other,
- * `tsc` fails.
- */
-export interface ZCashNativeClient {
-  readonly grpcUrl: string;
-  readonly network: string;
-  getChainTip(): Promise<number>;
-  /** @param timestamp - Unix timestamp in **seconds** (not milliseconds). */
-  findBlockHeight(timestamp: number): Promise<number>;
-  estimatedSyncTime(totalBlocks: number): Promise<(processedBlocks: number) => SyncEstimatedTime>;
-  syncShielded(args: SyncShieldedArgs): Observable<ShieldedSyncResult>;
-}
-
-export type ZcashSyncState = "disabled" | "ready" | "running" | "stopped" | "complete" | "outdated";
-
-export type DecryptedOutputRaw = {
-  memo: string;
-  transfer_type: string;
-  amount: string; // zatoshis
-};
-
-export type DecryptedOutput = {
-  memo: string;
-  transfer_type: string;
-  amount: BigNumber; // zatoshis
-};
-
-export type DecryptedTransaction = {
-  orchard_outputs: DecryptedOutput[];
-  sapling_outputs: DecryptedOutput[];
-};
-
-export type ShieldedTransaction = {
-  id: string;
-  hex: string;
-  blockHeight: number;
-  blockHash: string;
-  timestamp: number;
-  fee: BigNumber; // zatoshis
-  decryptedData?: DecryptedTransaction;
-};
-
-export type ZcashPrivateInfo = {
-  saplingBalance: BigNumber;
-  orchardBalance: BigNumber;
-  syncState: ZcashSyncState;
-  progress: number;
-  estimatedTimeRemaining: SyncEstimatedTime;
-  ufvk: string | null;
-  birthday: string | null;
-  lastSyncTimestamp: number | null;
-  lastProcessedBlock: number | null;
-  transactions: ShieldedTransaction[];
-};
-
-export type ZcashPrivateInfoRaw = {
-  orchardBalance: string;
-  saplingBalance: string;
-  syncState: string;
-  progress: number;
-  estimatedTimeRemaining: SyncEstimatedTime;
-  ufvk: string | null;
-  birthday: string | null;
-  lastSyncTimestamp: number | null;
-  lastProcessedBlock: number | null;
-  transactions: ShieldedTransactionRaw[];
-};
-
-export type ShieldedTransactionRaw = {
-  id: string;
-  hex: string;
-  blockHeight: number;
-  blockHash: string;
-  timestamp: number;
-  fee: string; // zatoshis
-  decryptedData?: {
-    orchard_outputs: DecryptedOutputRaw[];
-    sapling_outputs: DecryptedOutputRaw[];
-  };
-};
-
-export type ShieldedSyncResult = {
-  processedBlocks: number;
-  remainingBlocks: number;
-  lastProcessedBlock?: number;
-  transactions: ShieldedTransaction[];
-};
-
-/**
- * IPC-safe variant of {@link ShieldedSyncResult}.
- *
- * `BigNumber` does not survive `structuredClone` (used by Electron IPC and
- * UtilityProcess `parentPort.postMessage`), so the native engine emits this
- * raw shape; the renderer client rehydrates it into {@link ShieldedSyncResult}
- * with real `BigNumber` instances.
- */
-export type ShieldedSyncResultRaw = {
-  processedBlocks: number;
-  remainingBlocks: number;
-  lastProcessedBlock?: number;
-  transactions: ShieldedTransactionRaw[];
-};
-
-export type SyncEstimatedTime = {
-  hours: number;
-  minutes: number;
-};
-
-export const ZCASH_SHIELDED_TX_IN_TYPES = ["SHIELDED_TX_SAPLING_IN", "SHIELDED_TX_ORCHARD_IN"];
-
-export const ZCASH_SHIELDED_TX_OUT_TYPES = ["SHIELDED_TX_SAPLING_OUT", "SHIELDED_TX_ORCHARD_OUT"];
-
-export const ZCASH_SHIELDED_TX_TYPES = [
-  ...ZCASH_SHIELDED_TX_IN_TYPES,
-  ...ZCASH_SHIELDED_TX_OUT_TYPES,
-  "SHIELDED_TX_INTERNAL",
-];
