--- libs/coin-modules/zcash-shielded/tests/testAccounts.ts
@@ -1,234 +1,0 @@
-import BigNumber from "bignumber.js";
-
-export const LAST_BLOCK_COUNT = 3697079;
-export const testAccount1 = {
-  viewingKey:
-    "uviewtest1mqju2lk34ly4tt0xdtdsw0ngrzmx5prs4m09mzn0hq9deq5mcyngpxwzptakch8ays87pq58ny0exvzgns3v5lryxe7xj2fyearv8tujeqac0q3xc7pchmfn40lqwscvyqulqcgux4qlvep5utv5tg8kyhr9p0yxccvaxl6eklxprmga5a99ars3lnwkh04f3zumv9x2ulf5wpfawx2fxlp20kwckjgdp3t0sk4puxuh2t8jhrgtnaq4kue2qf429shywt5legg6kl8gcnmqura5f7dk3u00530wmufuumax5rp9907xde04cz8s3errjzlkywm229k33qhwhf7393mg5xkk4gejwdxxkv0mtjhmq3q0wztzdc7wtsnn4pvgf67q8xs6jv7x9rzr2f7qnv6teplzt052pa2nltfe6t7hm5r33wgplumhz57xgv0jwmlg9kqzrygxusg8efzx6xzw7xll2328kagq9sg0048p9hdpuqvgxu35",
-};
-
-export const testAccount2 = {
-  viewingKey:
-    "uviewtest174cgmrca27ujfymeppxdh7frr4y3a85g0zy0rnt5j547r9y7ylxakre3fn05zu9smrlmlsf2pnv3g2t294hm5ndd0kwuy3wa6nsn356wv0gx9ykwffhk9mf7xkr9e026ak7yl75ju0cf39njh8mr50dg69886vnmyz6gnjj6fk9nu9fx7w3xgll32wrwhtwfxhmld0v9uy4k3uv5nk4rwdq9fgt9h07chw330lh357ez90u5f58f65dx088l8s9m29xqnctc2expavj47fczstdl2tfwtpzgfnt3q9k73ek4xv0vwjth0xepsve6em7a44udvqh78d3nvp76wxl5hf7e23hzflhgrvlfhh3amzhfh3mamrvyjr2dmqgzu2v7w53dcmk7zndl3w08vf8h30t4mtr5jx54zyjsqh0mpvtf7flja2tc2svweps3k38mng3sg7lrp85enkgng8x48cw48wexta2cgqr3kvqmdqs4ukj0fsewq59m",
-};
-
-export const blocks = [
-  // block # 3697074
-  {
-    hash: "00267669fcb12b785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697074,
-    time: 1764164743,
-    tx: [
-      // 0: Not shielded
-      "43887332d4f5b19b050539e7165a5242a2e17cad527cca0756de73a4bb426a77",
-      // 1: Shielded - matches testAccount1.viewingKey
-      "85538ee4fe674c708e5328c7cea3d7a8aea82b4c8d5e0c8e52af4f5de12deacb",
-    ],
-  },
-  // block # 3795067
-  {
-    hash: "005290002b3198cd350add1abed3d0c53f736f5ed54e3ec463c5d017e602630e",
-    height: 3795067,
-    time: 1768842685,
-    tx: [
-      // 0: Not shielded
-      "4aacee85ef59aa1c14e1d3a6f22579e3d8f39b6d462d1381f7268a380b7c7224",
-      // 1: Shielded - does not match testAccount1.viewingKey
-      "0f0c8d2d9dd2eb767316075f04679e73e05ef352251bc66fc251aabada5716b1",
-    ],
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3000000,
-    time: 1764164743,
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697075,
-    time: 1764164743,
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697076,
-    time: 1764164743,
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697077,
-    time: 1764164743,
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697078,
-    time: 1764164743,
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697079,
-    time: 1764164743,
-  },
-  // edge case: block without tx
-  {
-    hash: "edgecase00000785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-    height: 3697080,
-    time: 1764164743,
-  },
-];
-
-export const transactions = [
-  // block # 3697074,
-  // 0: Not shielded
-  {
-    txid: "43887332d4f5b19b050539e7165a5242a2e17cad527cca0756de73a4bb426a77",
-    hex: "050000800a27a726f04dec4d00000000b2693800010000000000000000000000000000000000000000000000000000000000000000ffffffff0603b269380116ffffffff0250807307000000001976a914d843f9aecfc3f8f4327b9483fdd2b4e77873203288ac20bcbe000000000017a9147a86d6c7eb12ce0aa309d7391a6f338eba3c242b87000000",
-    orchard: {
-      actions: [],
-      valueBalance: 0.0,
-      valueBalanceZat: 0,
-    },
-  },
-  // 1: Shielded Orchard - matches testAccount1.viewingKey
-  {
-    txid: "85538ee4fe674c708e5328c7cea3d7a8aea82b4c8d5e0c8e52af4f5de12deacb",
-    hex: "050000800a27a726f04dec4d00000000d769380000000000025e8cf18333e28a70d466e5eddeba0986fa29d016f1471f1c7dee1114b29e811c5d957033fd6e49f25a6d1ccadeef13de02024b8fded34ea8b9c272365c2bcf3e2be5ad48ecac4cedbffe1b048de04a587b1087a6b160b30bb9cd83cd5a257a878ee2da808ea93a333b76b73946a5dd577d7df3fb5a158cc4ece62e9f09adf1179045fbad57c5e309d5d4ee850dc557ae5c2890a95f228f20abbd8a0d861edb073a2537cc5021c5303a49589d7f1e3234e42e5a5b6c77ba2de1d479b09d6a2d881d4bb2a49e5c7d1a59ed53d955e5a1317d26fa954f211442a619b940310552418ddaa3bd4c645bd8f7ca80b2064d40dcfcf155ab9c022931b666697c97d22c1ddfb48f0477b752d194fe87a1c9fc9b54dcdd63eb4dbe4a67efc88471060223d2487c18acb199c00d6459f6bbcba760465c19f8a8fec4e9c6096fa37dc167c42211d5f14e954f9e433704500919a14eaa973e8d3f5844c17efe836f2981a7820a509f003ba84d58d80c1607eb5bcf19dea7e9368648e1114fa10feb74229981d27153debe916730eb67f51f7c2ed18d605afcb138ff2cfe56d1860a916ec1509aa8e7353a3fff389845ce2d51f4c7b81dffba2a0785cc6cffd6933c6da3afe584b1a05c7df5aff4249cfeddd8cccaebc1b9b7aa36b0c7cc24c1b6eda843dbb76696242c6e297e158dafb7ee6e8bb74c4fddd953fc534e6bbe6edde479850c51bbc87728caf8d5e1ef744be234a41da0a93540250df172a5cff51ee681f13f45f355e2c869341ddda6ae643509896b7191848f081164627a4d3dd0bf4018041a3693e2221b74317eb599f27abdf711c64130b2320b4e4ac228db846abd16a790351c0e865a3592d11d376360ec387a2172f53c52d571fa13e6ef81730122d2ee85c1630bd49764e803a480c69ebf58189d0afb7fcb1e4760cae08fddf2a09340e4603cc7c82613c6d74898367cec212f282a34222f9a7a240eab15a37d1ad5acb016b27cd2bb5ac738584c3e83a91f293490114b2c9efceedd59aaa33d5ae17cbf6df75881b574b13c8c6e38da4bab0da2149229b6465120bb013fe960a18a0e88b5b3da06d4a9f3befa070c28de1713ed74af3c976a3855df054198a62497d8da8f393232a1cbe6fe258182d67424b909892fef112f73bb539b8b39b1cfb8b1540830a8484790e0b1ba4d195d6dac331d840c7f328caad310489cac303262e2c0065bce8d40fe1819cd43ccb02320612582eab0343df437519f30578e964e2b5602b5c21fb5cbdf5d6e96eadd9b0fe3cef2de05ade163cdc36722111179e5735738125b270b5d7653119fe475ee2814ec49c9ae2e32724d53b4704081ae1ad42b3db29bf59d75f02c4480544ef850ada17115333d94950a5ffdd9353d5a0eb9aeecd7f6811cf84d9b7caed1d166b46b63ffd5ac69525fad315be69a1db36f675422429bee3d9ae30bac8a3eb267262a24f2524f2429cb4bf2e7a1d261a15eda4136b198a4d06b0a4790d787b837a734edf183d73ccf0ff0019532038f7cc5e9c6283407e28282fbf34ea4e582b23bc5aaf20fc251ea64da39e305819f4ae44dcdd6ef3fe4be76afacf78b4548f74ce9f05f237ebe1adc5c4cc92b309cf541f0470cdaeb6fd6034afffe9b28387256637b7dd65b730ec65fe1a3094b5f17444211ab9aa3a54eca029e7697e859177dd41b52a3bbd711615dc15bc038653b6253ce50fc1aaff580fa2309183fa66f36529afb7b231b569f52aa1d7a8f49ff502ce51802bdaf7b692b63d1f06fbe1484c12fb45778f7df82c65d2c0bc058a0825fd2811eef836739414d12b80ad224f3d66f7a5c13b68375b3e6261d00e721f5055616e14fb424bd4b2fa40d1987e02c2db433a1e91ca0381f42c6b6f2b47957d1acf4debad9bbf86c526da17bc503fe8027b6d674ec3438eaed3a01ac02d51e2160e871171cd914e7f33d80a9762364209415fde62d35d440dbc31a40b0d9d94d030acef23745b50c07cfa2a2cee440cd857eb01a3021f24855395b6a1307bc2f3aad6e011462a3894c6ff0d3fcb531a0ba6ff6be66c5383b5b220e96489b859f76bfc33707132f5574d5a288fef43f7774a31f3cb97b0cea7b0b71e8ec9b2458e74d3687913de1f86a86f2ab10a677b39029d1b26040b029ae95667b316b200451fe100d423bac33dbbbfbe3d50f04d5a660b7c00f6b7b8f40442e5549f0c4ed223a662460e2eaf2657d857514affd3cb77fc1f462f779cac6d03350e741cc19186daedc84d33ac22700261ef957e6b65ee01f3c68e4f0f7ef526fe109cc220b857ba2347272a6f96203102700000000000034281bf409fc8d9c4f12be7da62bc8dae72c00072e35eaccb78066b7cc801b15fd601ccaac79d3f9175531c771dfca7fe28ebda25b333ea6dd92e9e240fbb743c6102c189ccbf8be16980199cd300dbdf14c0cb053e86f0780a020620bb2119ea35525b03056cee69971e2e27b0b67a3a189de7ff2467fef145b06f44dc710efa95bbdc95b6489af9f01295dbd8194a33c6fc3a4570130f7d55ab7c22cc8abc714963d660056282311ec3ff6a0bd1a23372154f553e8f9c0ca47b4a81b54fbcb7a2787031d5ef1619e6f22a2d410e3d72a1ca48baaec5d19d473de228d07731e3dc1039551d9f884850194540e8ed87d843c6875bb898bb3517f6c218c8b3e1069812df11a46dbaa7aaa08e9ade17c7d9ecb121505315cf9c5bf4dc9b0ba77ad477db6c57d006da7c2b68e48774b774d9f6c95d51e71547a0c72161f64fdd357720f2955316cff36c990dbe18087851b3a3732b16accb90200cf9eab0ae1f5b09814bb1bfcc47044f6b284c82998924cde9eeb506f990623115aa790a64df7b7743b9b9ca105f5d0c71780d73419796f8f354c8ac5fa054a2e2716dcc62f9bf2613da8ff4013395c0edb1ea68ab9051f17ab3f5a9b6885099c2b8348eca9f724501e3077fae01aa0fbdcf78cdd4b4c85a24fd3db2d62a049505ca5167de472a5980fa3618aa251c1aed6eca6f2fe75a6b301672669f99cf1dbf9c9b415a2e3f54cbaaa62b9486d4868b6bc2e023dfd2d6c469e27aeb2d10663156a0cfc8df94743f53acd7a0a5ecea67c424e03d80d86a93432bde6642a5381d5d22f760345145b901b2b44d6020c0c2bf4a6b911917e11f4abff7abd00be30f48b7abbffb21f7e01363157ff619bacb119b98257c5fc5bde966f66c6462b9c496c58b33061de169389ac3fb4f1d61c2440dff1a8d9089cdfa61024e06f8ce7077ff250be93ca4738862edb599236adcb336f163e74a2d3655c4385fb49b4aaf12d80624f20a249eb0054d588ce9164c6dc759b97434ec272e70c3fd6fd87d9325962b5ea020ab9f582cd44da34e90681c22e666ef4c3f544a756b0c5ae2fa7e3d676ed5df4bfb7171c0fbef9f293167be02ada2a82bacbb3671900435c8d2e5b2ea20b71bb6a865629940034fc398494405bc123471250dfa9a9309823fddd1bc69833d833b412de06749622e163dc93f9c8f24b33864be8cc8b967fe859ca89a888b12cb6f4a085948dd39dbe7e158527050b5770f39c0c154b1bd282c017897963adafd1bc5ba13da4d18a083fd93dcc0911571b46c06c79dbddc3d2b74c43430982f4573d4c1e8d0f87efcd190e55477de89b06b928fe03a62fbda36f85c190b44df7cd9113f52a9a6ebb365ab6f2822042f8e46513b3cab44691722c8fa2ee14c1ad1e93a0aa17cf706a26e6b40ca2142c9e6103c17bed1d54d20a388113fbec092e376adbf60198b75bda9b48231cd50b118ef0c29a2fde63a65a4453ccf46709243a0a746aa8ba74aeedee37cf92cba1b43bac06e705e01dcc1dff0e7b422d86fb3c35bd6634be237b09e8ce8ce181c8b4005586f999770d68adfb21201bbce2cb5a7221f02ff48efd7eaa8b35a0aa79ad4ee122b6a695a8b098f41411be905d04133466af8484297b24f0a6e7adeca1ef14019be9e20c0a56cc8f4943da54c80934d36e0916f7d1fd20f87d656466a9d619a7f57608e9f3143d5f8af1dff768f0788681c32090affd90ad56da2a1da6c5a662e2f0860b3825c38accaafb91f20041a2c63992a1d0024cfb65015ca15b930799412a67d6c29f462f53dba7da6ed1ca6ba5af3712f74416d7b195d28e29fe8af3ac0845869c220b00369a0ff012d606ce71220ef28632a792ed19a11fe0600cb277f378f7c2e6ef56df37cf6c3af44cd881afb620a7cda6e56da74803c1f00ef5614d0ff86f4927dbe61f72fc859fa66a22209f5c67d1a452c073cbfc6047121b2bb20e032142b0df5750aca056ffdf340ab13592f54ffb7d9a004f5b63656fcf4d2ddd42faa1371625d9ef38cd9c92bf815337477bfd25b5796ef88587717d4ff11fee696901905f4026a508a8c65f0e246c36d3ce4c8598d5fbd7d040d86aa222d1ee92e2a8c81bd309a688e637892f532c221a91863f0684982399e54475a0ac9a13187abced6e02f5c0dd2c82933c76668f986c4200c65ef2ac8ab939730eab61c87f05a0792374c69aa1acdb8b45028d8200f45063aebd86dc9384480bb9b376a1528c885af2d8d9d10224231612cd472d483e4c980859a4c09db035c3d091dd4122755ec0beed9635151cdb17adc006b1311896f46365b18509d9f9f5bd7afae39326d9c2c02c4930e0e40f3236c92cb0a79d3de17dc9ee52ce3b598b91a91be6cc5b1dfa28464740a9cfa3c942129a3b119e21cedec048370a6cee6a5fbbbe2fde6395762255e10da9779fe6966a559999b5bb90e56d467acd95dfc9298b2dd550f0d3909f404492ac161d1209faf224d90aeae1d190cecae3f3fb9b9d215d71cdf07b77fda3e7457b047cb32f7f541a5f1dbd7ca739c30a981fe0b1805b1af2deaa4d346d15cfe00e22bee751724d1aa398a7787e6a3a2fbf7a0c4d380ada1e9c5e126f74841ad3c8e144e2f6911a23fc2ed971fc6886d7a294cc883b3c92f531a4ed9992fa9417e92a3f5d04efe7333bf9b3bcc9a1a253726cbf5d05f02aca32ba6cf464cc4c9291349e483c13f821ad6da0641cdbcce20a94bc9185fc64002aaf6546170e166cb5adae2d887d5100bc065bb7c05de53431c0f5b3cd665366d118240071073303d295a35e4e06a922dbcd7cb6ceb2a8ed4423b5990168ca7523dc4dc164d22c9f78600ba02e6f0e1964df7197e8f581ef6d82d4113007600d4858500a6793a477b3230b4fc3bff10d38e5ebb13592ff4ee6aca3b841f006ef8bf90e33809039c9fc97479dbb9d2223a542962a730d0518b62f638cf12e37c1fc80e48aa1fa2d3add5438a0b07d1d3f88af2878021249cbd1c3b2c750c9c15dd7e7b17aefa8e1cafd8e3f3c9eb3e3374896d1074db31aad601f18cacafc99692f7ef872455826e820056af54c0dbd2c2688882712e4408f6393b363729ea115fd22f4ce663ea0c9f5e5e5a448df92064ebbd8e987bda450e9a8b1dd6bb9c95802d5219335107453baf035fe079271098c956f4080e1df98911d05f3b1f2ee33a5d51fb01fec589418700caf19d363352f71ce8fd97dfe8291e05cfa0f23c016e97c8463598ca68871f84512ecd8373d46894f762b79ccd23d53cb72b00dd74aa60a3732d41eb62155540a366ae9d73edd4fa47a61340df7197ef1416939126ac3afb62fe5ff8aff82d4779c65a1d0226ada87a55cc886c7773229f262c943f206fbb36c52eae3d1b9e53d725bd4d81c2673df972a8c271011339b11c3ec72ea950454e2ff7c100ef59e5c3d01ba142d02af71690a0a12c0c8b3a77e11030284340a74a99ad8f0d5418ded1a1e8f1612061e1a0ebf8781175ebc71a8666c28aa1394302df122f86c2353e5c3c7eb6b3969030842c6538434f2055aea9a32157037456f23e0bd1336c8e5dfeb119c8012d4009211c5cc782c144ecf255f40e35da540f84877fbad423267f1b361df5b22b00dec86dd7eb78b5b22ada1e64537950553c54483544db329fb720071da33314988466ec055bce92b16ef450cef67e47c4dc1e11e626463dd517f4d490d3e3e4e618bf05c14ead979d652c9e20bda042b719ec67c55d96271b5d62476000f37957933466f593e5eec0f716e7e01ce75415571d8087d4cb89f54717845c81438717d3c4e17e415c772d54533435502355a2c1d71cc8292a4c8079e87e5c31e139545cbe7acf583dc6ad5b6a72cbe03cceaf36d5a587fbe7e9b68dfecbdcc193a0d8d996a9da7b951ab374305e88a18c830e2463cb44b78dc4a76547f8997a11c8a7a000d6660e2e6e540330b6c979ee4b1645e8109b2dc1abaceee4374b6f9006ea2c99585b5378050a76718c4223f3b06893818b10bbcfe82faa95e96723f15a5ac1d0d472d57ede2708242e66d43c589dabb19bd1764943f111d064d14bd0aedb27fc4b427b3bc2d76c105341df80d66ad0ac8179bdd4200305cccb8612b0812034813c792b9aae9e6c3b1f008b0932e5acbd15d795355eebcf887afdfae12535b0f10c8afd2369d8c7f51bbcea37283c2005decab417dfbfe85019224f11bed795a4afd4c2c4436a5da0be0e4a73f86f9327cc24050ab82a398920cb8ca39ae1d0da5aa93679c21f0536e6d7d26a043fee771ac5ac429f834049646131b2b5357162ec017e655b035ef5701b7bad42004eb862091678ead9a1b6409e9fe2c92cc6bd7d291bbdbef9c2b5daca811d2028e87df8bea7dda0f5fb377040dd6233902fd7fe4605233fd5f517b83c01d3d8781a32d94c930f88ce8bc90914c833c3727eb8d745bcf878126617d798be3443189e8d36a5ecc969d5c6694c0403f27f1928fb1b639b52e88b028aed064541e307fba76e0292dda3cb15ff0d3767a129b7bd1e7040dd828583da010b68e77f830ba82f66628bd89013d8a632c9f9f0e915fa71c634b87b50129836d14dccb97a6f97d661d13f117953e267824b1d4206155e404f61695f239faa706d0667293c1c554cdc60791204227fbba6c903234c1ca571d88cfb9826720f7fe5bf002c71e6215c871b0462d36c20fd7317b453a7e201d4997896c099e7a68a08aaa7e9cc3e6a2726985749d68f9fca08e147b22c5a6bccc62376bfd2781d599d050aded284340e03529dbf7a854ed7c4472e739af9d75d0b84425eeac563b197a0bb6aead97fc244fb1b0156c3003ed9971250c0cfe69b34c0bd6161943d917e28f7fd31fb44cc5adca98a64a1c13154cb86c357a0fbe0120bb5c147079e7271616da150cc9abaf06e743ad9d58c734337cd12fa1ce18b2944a42cb15f0aaec830937831cc3e07533470495e90efbab58434930ff28f7f5f0ab2e748d7ce36181da92800b792c4e53689c853f850610821cef1074824b2b5e88f557911ab0dbd87387185566131a97f7d5e07519f6cc8997d40d0b6422031efe083b76605f99d499583b374c3acc0e7216e8ddbed71fcc2b313d5fb965fe506e87ae651468b74b7b739f201962182e65213b96081d18596b7c35a61d3968b08b11adff278dade01982dd929c2f44f4d984c227a0a2371de9e02d53eaefa31da0f90c2d24a7f85ab38edae071326542fa0a81251d4df241b49600499bbb7945d617872cd55021669cb8d2f3764d33e0249f1160ee55a750557b2fcaea8a76f341eb05aa35933e8ede9e7371c89ff87acbaa2a24d7ed0000d0753484ae75b11115b2e2d32166e33ce88b7f454088ede8b4d580f48b322724cdff37834d8d157cb073238451c03b74c331eb12f13d4487e75b3f36b6eb799df14f23593f0c8b0ad1bf5a0ceb495fcb0c5667ede2adcb789857611b13b324c0ab2e18a294e9adc6313d94c85f550e24f7a56f8b4179be018ca6fa731d4fa62305a02e115224d78641659cfb691fbc80055388b2de4afd75538f0ba085c4a734e4902e3978306ee876775e9d6f4bc78d593a8abf8cdf1ace2c9b849e1327c5844963373ad2bc99f460be677edce5bb1b525d4b05f2964cddcc1c3a8861ed75a3738e0a5f68a22cc4196c2a471c52a9a1da756bdf7abd069756f2fadc9db4d6b1f5d52a28ab54ea166239605d8e0d5ad01a027b9006735a58c9239a1f0b1d7594285f2fb36167d5a84a3d8a80d8b0f359974057fb81ee5f4cd466c29679564dd87f0017aa77412c68bc662fbb66bbbf2e6da431e634852ab0196ec8f492c48f66f86f10ab22a581ee16534d49c6f4e108deba666ec59a718a72fab1adba773452c3ba1e2e0f96831456de24abb41432e649a72cfcddbaec3b95950c2acdb25cd3a7640d4a243529cb27f1bb141c3c3bf7eaddb58c6451892604a6177214c0387c71f62a39d274d85539cf3028d61e18df864290dde5d860536c769275c94ddb4461492a8880940ad4f9b25b4626ea475e1ac27ff8040ee76f739cbe0fc4c78c3311a41fcc913dc7cecdd5a5e6cceb0f6c742965cc059751ae0ae888ce58356afad72d271b9e0148adf7208d3d473d195bd708289140a017b603ef7dacf3a8d1dcf88f397ed00690ca1f4d6cc08d85f70e7d1d68059023bc1182f3dda2c84f6e927e7f0ac8e9e87503f61c557705a392c6409223ebe9faeb400ec7cf5f1e629d5068610ba0dc096f0ea64ce976beacfc58a6febb6cf5855611b9dc3ff3f7fde67d479539192544b274c6266095657d456e7f9dc3a7811ffb79dd98d8446a05181fbd8a13e069139ef99c9a921fa16e81b13c8c02e4f5ccf60c03d162711056e6a29ff701fca92edff778d16a5206731de6009da44d5a618362f145cd644b63929ac1ab14d06b3af3859301169c04aaf79992ca41aada60e1b415db6ec6cef1b1104ea809074a63c4ef8f1315b84fe62c24d8f740a8c279f3b7a59d50a08185a0733b891f75af18dded07186a0d19711c717a07515c6bbe409b9cb022a10ccd25c70c463ce41913881b04f7d9ac23492c483d0655d91b560d623f69bd99a878c945794522070ba66b660c1e6e243780b0a4e74bac031cdca4e4919a19839fa810ebd2a92299ffbfc75471db02736b19a54d45fe3291d3f8189734bd1a5791b97e538bb136c9db806a5f52aba81de6991fed274f161b1fb2569b5683e2b54a57dfaf14d40830ba41b99bef7bf958d1574dc41969b2adc9f360a71f1b3dc7ef6b2bbed741308cb4be11d49b97cec69b5c0f96ad760f7ad2c364595a6967991b6e15f32de52cf5086a4361dcb517080b2da382ff7613c2f153205229ad4a67754490ae79bf121d34f40e33efc114047d99dcceeef410ff072673604a2a4a39cfe5989ef202202946df3584c567ac298915f1452eda61f255a2d50bfa347e9f75b3c9c417a13cf4ffee5d163e77dafb6bdd5ec63930880ad4b5db5efea83fa98d582147a5ce1f34679682dde6adfd94d925b549552c5f9f34603dc11f211d5698065d69ecfe1e8ce4ca6d83cd94e631fe3ad39153e860bac8899246ecd8f0d22a2819a3a4513a250593cd6db24a8bfcb7696b6870334dbd56d35cbc3c6d98baeb33f8d2456d280346d79924d61f8fd614f4ecc4b3a07c6fda10d9d5f0e7734b49883753174e35d6338545dd8d39b28f948007012ecd577739cbca9068c8a38fc3e7b1ea97440dc251c48c5d28d40f938ec65d20679a7f725009807acd4c4f42228989e009d826f21b8b7500cda986c94be14d5db3cd339f3711747bd2a873a1a5d3424b8e582018a2cb24707b6adfb1364abdf761205fcac6436f99ba4423f06c38ac50fd6636ed4ffe0c64b6bf8bf21c8e05b100be9fbdba7d67278c7dde7ae719422476042d24b7168eb25ec8d324636beba396323083691eecf1465e557c7c52895bcbae0b376ac12a566bdd32b0f9bde8e8807df5f0013d13497c5f7476a1cdd56e19ab3ab68679fc886156b4c1964b1f132eacab0a3e4d299ca545db62c833a9c7117b27f09e375363a70a5d5427a38130f3287d47b30d9c946c2402b6a6befa7ac7d923c38fa112f51b8a92f2f48967a6d07fbfbc35dca10221a540208aa9b7b1610b1fb647c23304becea00357a06cb371bdeb90a1ed3976990097b640b6e1511aa2349f4a242f087ff5192c4259067ce0b998cf14780b33fcf0ca8881fc2e6c6824390567a6bcf1101fd68ec3e30c5a8d536132bbdc4f0b23022c964a3d9bc563e3306806db26a1263247b54c1c53ce556eb0a4a1945a44e6f1147834258794dba0153c4847dca0d3ee2797aa2708923cb1ed513f0faa39d38fa6864b6aa83c69e915f351229b33a091bc1b4a43405adbd9e51ceeaecc081b625e2124be9b8c11ad0f5c6716b8aed58da92ea060954ed7fbfd8da1c94bea2773d1fb2c2c329bcca51bc122c7dc704c1ea30b062bf0fb98804a7e4d0d88a0f13d2441c1d0f010eaa2143fdb3ba944fe4b881fa3659cdf887864307213fa1c6d86d4fb0595c68000fb345ca45b77057bc943c1647f8378004e9157e63034777e6208a6c4d33eeb4ab20de05daf0bc0f877cab2f6333f767b9460f81df870836ddea63cca323d38d5f60533167d45360dcb2ef6e8b8931e6cc57188fd909164390877e988974d39166611cc97399fdb2ecb607b719764862a773e022376ccd319f995773cb3d7ab4a612c50f619f3d34ab8ddf35f310d06850f7d06daaf8ddedeac096cd0743b329a08173933a9a08fac91ac818e73a2ebe9611390186a159af93ce5fb49a42a5826ec2cdcac3c44047feb628454e4bfdacb089811d96c6837be30fb0d2a9f0d514b5d06ca174bf6b7792db5e4d4d40833a43646f9d8c695e12bebd4f79448a0d3faeb16be334830e990defe412faba0c978a3fbd4cf2fcbf24a5a3434f5cec707adc32f9b143c2d9fc09831620bb434f47b2c2e34f9f0fac64f26041d70ddbcfde3421db4ba1c845c7e8289e95bdcfbb7a6ec98dc9b022000c86ae581ed8be4bb93e62e69e2e6147b386181090c36b4710facf80322ec30ebd62167b867d0162f607107bf9f8ebc1994f122b00a2175f19aa5f53c3c0c2300944a8945dbfc04a01f04350737d9a8e48139cef503952315c99de10341a40ca9a0da32f2d22f171ebe35102733c4763ef2233b41e3f51adf7cb5b60633496774988d315a466270efac7a1d1e560efd15461d3107a4a813540ca4ca3c2529fc88dc29a0a7016062ca83ad2c2893a466e0af9d511b9a431d71caa0045fe2c0b1ec054b6392627ab98b21cd382cd764ac0592c4e26f71c8f1b36121783f0ae4bbb4d73b2f680102bda899d80c157d1270778733b54a1d9b7a9bd5d0d0146683e7e2998e6cf9b69e5c1c6e791c7728bb10646fd51d1502d1b852fa48cf10274f30cb9d676304d0a9452a554322b37dd5e5eed4d8fcc54dffd5af0296de91e0f3e731fe9ac32f87f0d5fb43152da318e0cc2090176d730a2f15e2713b040c0bff87df75b2e969dcef8a4edcdb392d6f22c3ab5392d5c22f0568bb27174290465fc9d2a8e5d4933a4d33089d183416d1ffff729d92ef7dc3fd63bd436d41524c5c802ffa12209ff712915b8962256eb7a470cbfaf473cdb2c153db7589746d506c753f8597bd86f53081a3fffc3e3cdc2d2275c72b6fbaeaf2900e4f3e325211ea158b1bce0cfe3d4e861053c42496787a9ecc3f4a2e013fa3636e4071973ce1db34357dfd2fb8397cabbc772b85766f06c9beccf71c296ace51903d718ed68a6d5a04b41d5bd3bba205e187a59c8bca661b1a72bf8892bd756c87e0d7692adcf350882e47d53dca6cdfad0a6887ea548d61a665d4f77dfcb275cb26644fb1aa298b7f685ff0fa59d1567bf9a19b95a3dcf8a3c3d0823e1a0f7e15256f9c8feb7290407231e4585776d19628001f54c6ec36619ed5383b09f83a0e5dcc6962c54ad6bb2dc4da285ca170f67bfb126d9f5019c90a578a50ade6e60397a16e10aa6a088c8189fc9b0969150f2e1835ae7e7bed51956b3feb17213546a21ed45e9e5ed554be8fd9fe2707eed5ddc11afa50f678196cfa0dc5cbc59fe43fcb78e94f9d07d4ac49520142c27e9e49380ec19beb993c7f7d1d80d97541e8dc453183c6d749c5d80683470dfc4f33be018d7da0941c1e5952b0149f5cdf898ff1a77082b17fb2da8e052997f3020ec612bfbc0287f41f2dbff37b334edddf7936cb5a7c4e8d7046d7eb98505dbd1b47fe1180b8e259b004c91f3e059ce895024342e16e2bf2095c880552e2116468a2861e2b992d39ca7e084c6afcbe6f493a42381bae4a5bcb7c16dcca4f72f3084335b0c2308dd58461a0ededa477d30f8a239ddfeefa889eb28481731ac703953acc87d5d6235da72c9ec0ee24172035b0cb2d40ae6b745cb33229288c48dc1149bb05488fb4afd590c61df95a2cfb90cf393582020b2dc64a799c1d1c5069430e953e5e2fd15577fbb94fbf96db250a4aa3682160c68e33a30549a41c6d571e51460e07757716211a8fe1c76fd029907997672f4bb17dfb550b3e4fd616841e19123713f94b92090be9741df822ae08a417e15a1a67713b50ba4b40434001bdec3fa5c48c20c2bd5c1897227b52eef1084bbc7d19ef2910a4780662f1f4d016ab97af07af81df625c4c569eef1a948b3ce8a80f662784c5d9ce3ae85b90793aca35a7dabdbd9c13ea63f947c9c7c73f3be4ee7c3de2e1c39593265d817e63dd9874063b6498f83eeb6bb8af32915a4337c98cfe24c8a3603839da4b7d56f18a8892000600230768b2a9c137b521fe5ee686c1e473149dfa70976a74ec78029f02a939007642c0215fc4f4651fad63eee68fe2884a849c930147d7a0744f3a109ef91b5f09f67990ba5e5690e9dc1ec23d1f1e524eb70a0952034302cf6a0339f286bf8929a4e9a79d3a3bd17584b26c23425cc6239ee19f60a8206b8e3e388f827737c8cf04f52a843240313f28dd93ac5a4be2eb5398996bd8d2f494bf1ac47a188bfac3897e2249f06f0c39d9d45f31fa1fe3a660ecc3225255a71d927566703404",
-    orchard: {
-      actions: [
-        {
-          cv: "5e8cf18333e28a70d466e5eddeba0986fa29d016f1471f1c7dee1114b29e811c",
-          nullifier: "5d957033fd6e49f25a6d1ccadeef13de02024b8fded34ea8b9c272365c2bcf3e",
-          rk: "2be5ad48ecac4cedbffe1b048de04a587b1087a6b160b30bb9cd83cd5a257a87",
-          cmx: "8ee2da808ea93a333b76b73946a5dd577d7df3fb5a158cc4ece62e9f09adf117",
-          ephemeralKey: "9045fbad57c5e309d5d4ee850dc557ae5c2890a95f228f20abbd8a0d861edb07",
-          encCiphertext:
-            "3a2537cc5021c5303a49589d7f1e3234e42e5a5b6c77ba2de1d479b09d6a2d881d4bb2a49e5c7d1a59ed53d955e5a1317d26fa954f211442a619b940310552418ddaa3bd4c645bd8f7ca80b2064d40dcfcf155ab9c022931b666697c97d22c1ddfb48f0477b752d194fe87a1c9fc9b54dcdd63eb4dbe4a67efc88471060223d2487c18acb199c00d6459f6bbcba760465c19f8a8fec4e9c6096fa37dc167c42211d5f14e954f9e433704500919a14eaa973e8d3f5844c17efe836f2981a7820a509f003ba84d58d80c1607eb5bcf19dea7e9368648e1114fa10feb74229981d27153debe916730eb67f51f7c2ed18d605afcb138ff2cfe56d1860a916ec1509aa8e7353a3fff389845ce2d51f4c7b81dffba2a0785cc6cffd6933c6da3afe584b1a05c7df5aff4249cfeddd8cccaebc1b9b7aa36b0c7cc24c1b6eda843dbb76696242c6e297e158dafb7ee6e8bb74c4fddd953fc534e6bbe6edde479850c51bbc87728caf8d5e1ef744be234a41da0a93540250df172a5cff51ee681f13f45f355e2c869341ddda6ae643509896b7191848f081164627a4d3dd0bf4018041a3693e2221b74317eb599f27abdf711c64130b2320b4e4ac228db846abd16a790351c0e865a3592d11d376360ec387a2172f53c52d571fa13e6ef81730122d2ee85c1630bd49764e803a480c69ebf58189d0afb7fcb1e4760cae08fddf2a09340e4603cc7c82613c6d74898367cec212f282a34222f9a7a240eab15a37d1ad5acb016b27cd2bb5ac738584c3e83a91f293490114b2c9efceedd59aaa33d5ae17cbf6df75881",
-          spendAuthSig:
-            "0600230768b2a9c137b521fe5ee686c1e473149dfa70976a74ec78029f02a939007642c0215fc4f4651fad63eee68fe2884a849c930147d7a0744f3a109ef91b",
-          outCiphertext:
-            "b574b13c8c6e38da4bab0da2149229b6465120bb013fe960a18a0e88b5b3da06d4a9f3befa070c28de1713ed74af3c976a3855df054198a62497d8da8f393232a1cbe6fe258182d67424b909892fef11",
-        },
-        {
-          cv: "2f73bb539b8b39b1cfb8b1540830a8484790e0b1ba4d195d6dac331d840c7f32",
-          nullifier: "8caad310489cac303262e2c0065bce8d40fe1819cd43ccb02320612582eab034",
-          rk: "3df437519f30578e964e2b5602b5c21fb5cbdf5d6e96eadd9b0fe3cef2de05ad",
-          cmx: "e163cdc36722111179e5735738125b270b5d7653119fe475ee2814ec49c9ae2e",
-          ephemeralKey: "32724d53b4704081ae1ad42b3db29bf59d75f02c4480544ef850ada17115333d",
-          encCiphertext:
-            "94950a5ffdd9353d5a0eb9aeecd7f6811cf84d9b7caed1d166b46b63ffd5ac69525fad315be69a1db36f675422429bee3d9ae30bac8a3eb267262a24f2524f2429cb4bf2e7a1d261a15eda4136b198a4d06b0a4790d787b837a734edf183d73ccf0ff0019532038f7cc5e9c6283407e28282fbf34ea4e582b23bc5aaf20fc251ea64da39e305819f4ae44dcdd6ef3fe4be76afacf78b4548f74ce9f05f237ebe1adc5c4cc92b309cf541f0470cdaeb6fd6034afffe9b28387256637b7dd65b730ec65fe1a3094b5f17444211ab9aa3a54eca029e7697e859177dd41b52a3bbd711615dc15bc038653b6253ce50fc1aaff580fa2309183fa66f36529afb7b231b569f52aa1d7a8f49ff502ce51802bdaf7b692b63d1f06fbe1484c12fb45778f7df82c65d2c0bc058a0825fd2811eef836739414d12b80ad224f3d66f7a5c13b68375b3e6261d00e721f5055616e14fb424bd4b2fa40d1987e02c2db433a1e91ca0381f42c6b6f2b47957d1acf4debad9bbf86c526da17bc503fe8027b6d674ec3438eaed3a01ac02d51e2160e871171cd914e7f33d80a9762364209415fde62d35d440dbc31a40b0d9d94d030acef23745b50c07cfa2a2cee440cd857eb01a3021f24855395b6a1307bc2f3aad6e011462a3894c6ff0d3fcb531a0ba6ff6be66c5383b5b220e96489b859f76bfc33707132f5574d5a288fef43f7774a31f3cb97b0cea7b0b71e8ec9b2458e74d3687913de1f86a86f2ab10a677b39029d1b26040b029ae95667b316b200451fe100d423bac33dbbbfbe3d50f04d5a660b7c00f6b7b8f40",
-          spendAuthSig:
-            "5f09f67990ba5e5690e9dc1ec23d1f1e524eb70a0952034302cf6a0339f286bf8929a4e9a79d3a3bd17584b26c23425cc6239ee19f60a8206b8e3e388f827737",
-          outCiphertext:
-            "442e5549f0c4ed223a662460e2eaf2657d857514affd3cb77fc1f462f779cac6d03350e741cc19186daedc84d33ac22700261ef957e6b65ee01f3c68e4f0f7ef526fe109cc220b857ba2347272a6f962",
-        },
-      ],
-      valueBalance: 0.0001,
-      valueBalanceZat: 10000,
-      flags: {
-        enableOutputs: true,
-        enableSpends: true,
-      },
-      anchor: "34281bf409fc8d9c4f12be7da62bc8dae72c00072e35eaccb78066b7cc801b15",
-      proof: "...",
-      bindingSig: "...",
-    },
-    time: 1764164743,
-    height: 3697074,
-    blockhash: "00267669fcb12b785eda2359facdb9c49ddfcd43e1858a23d8051937cbda43fa",
-  },
-  // block # 3795067,
-  // 0: Not shielded
-  {
-    txid: "4aacee85ef59aa1c14e1d3a6f22579e3d8f39b6d462d1381f7268a380b7c7224",
-    hex: "050000800a27a726f04dec4d000000007be83900010000000000000000000000000000000000000000000000000000000000000000ffffffff06037be8390102ffffffff0250807307000000001976a914dd8aae20c7f358e8d21c2f32972d70ed98e0a81888ac20bcbe000000000017a9147a86d6c7eb12ce0aa309d7391a6f338eba3c242b87000000",
-    orchard: {
-      actions: [],
-      valueBalance: 0.0,
-      valueBalanceZat: 0,
-    },
-  },
-  // 1: Shielded - does not match testAccount1.viewingKey
-  {
-    txid: "0f0c8d2d9dd2eb767316075f04679e73e05ef352251bc66fc251aabada5716b1",
-    hex: "050000800a27a726f04dec4d00000000a2e83900000000000228192518b0209ec32d9d83cccecbcf5428c156c774f54878dca774b2fd0abf99bb2202837283f8b6346ce90ba21d9c9f3b3396e20643385fbe2302326b5487086d34f2c81cabd778580abe233e2b367a79fe2b23b1e127508a9a6fd23676ba8ffc2decae6f69f41d0df41d5a6db4f6c915f46962e5d4209b8ffe8df3e4fff2263d350536f4f3d826ed0bfcedccee16d737df585a56938b65e3587c32c486bdbf8dfb2a622832a346b79fde21144afe6436e1ea4e6520dc4f9988afa57a910be27f4614b3e6ddecba2794659fcfb89d9d976253d5e555b8c6db28d5ea833bf6f68e114bd80fa0eb14775f660e49693eb860659d97b1cae54e29f6598e6e40530a523d8c60e1f65dc265b96946abdd0d8844a5685f336d9a17c4fcb1f957697cc63b962ac27a655ce3532f62d9bdf17c2375d31f32aec69d71c8375ee9a1729ae3f081184cf55a639a659fbd5e26fbf9bda7f52ceb589bb39002b6fad43fad12d2a16ab0cdf34df6290bf69be7ac6d9512a181120f42d9fc7166f4880c3b2393a7de19f143a538bd03212e1f3bbe4dc2e143a166587f3b4fdbd4a0dab51cc71cf95d59866213f49703ca43db3fd69215ddf985c3bc31a08b4ab7b32f101b4dad87326839f148de0f59b3b9f820855ab649ceb0cd89cb27687a2991909fc5d2a4996ddad241e91c1fc53283c791c02a2d985f09ba47e4b6d46a5025b6fd0c335edab9a4f7562431d59c22368b46e20dfee70fed4e940f82e0eabbb5b21ccc005a3c3939cc55a3a362fd48efe5b8736fc2a6ea2ce942729a316eca35135b53ddb663b37d93ae64249a0890dca4ed86c72879cf248ba92430a8f3eb57611b7d4d5a89c9d517ea8cde7de7cd7f8316807298d2bcf732b0113a92912b87074371a1b4c353a996e15619bd813ab9ac41a695f1678c7cba0f8806fdf1d8c1d3e07d369f6036c9df71891a27f5f44f3a41b246a9ad5768cad155ace94ee732c2fc3d3162019f2d62295b2cfa4a62271446e9b7ab77af001dfc6b136e19dced51b5ff9e3f253877edb7e62b1a887ba55d7f1058d359929b93963fe232afdf56b92f1ba6d7d2a92300f4b45bf7c69366bd708edb56db4192748f368a05e76c002e9c3bd03ee91f92610a9310c311472a922a6ed4cfef760277c7bc162c43f73d6ca8ee4b8748e8e3473b237d15f78f85cd8969289a18f85a0f23be3f5f0eba8529f434a54e363ad7b20901dab8c7d92918122c1d9d066fcc970638527f3c5c67257910e55f30de035941712c10957afe2bec57c0644ee59bf93015bcff740c52d7d6647fe81cd8b332e5104b8bd0b0ce685a58004ad5b8279e2c53e654a951c5d7eb0a31fcb1e554907a435d00f6473d361cb076a9229ea416265e0a04002ee79238fb32d00529dc5198cf23cc9456023861247d8259369457e546697951141f3a3db58f327b191a765aeb5f8cbcb7e6b2024896a4ba3dbb462e008fe8bf320d52afd8cda355fe9a42ac9d3b11351022659b277ba59d204bd10d5b36532e8fa11a98d7ee6a9293b033417105be66e742785f3528343d987c33ce2e5c9dee2d6c49873ecbf8c54c10fd4cf27721004eae3cdbc61854a78aec2a1dbc552989dbf7dfeea583fa57860083f2611742c09114c7ceac786f4e4e25ddb4a6ce83f431efee6d1f8a3e6675e7185431b84056740ceb07b5ae2277d7efcb45739fc5cc90220ac59186294e466a1bf37b231e7965699cdbad494db0ffc3fc5c2518d412abdb6a80cc69b64504e1f29ada9dfc5fffd89a8ed524fa059bb8be6b12b0636371ff5dea026e7a17c1f1595d2eef4f876bfa33e63ffbe3f2677406121a661c0fc7c4f0bcba7a808f5d8c146bdccf436988ea521df47978b6db1dfcec67a5fc8b71ffafb97c8918d640ff37a66ac6763394481f5ac79c4a51a1a1038672b0e69b675b0503132689b57bb57e0891cbe25aacd6e6ada436fcc3da1003405f052571ea9d5da929a89e7b7f826b4ff622477a2abe897ba8622d5b4d221a6b2717058bf44fbbf04151ed9417fb5fcb2d381bb0b6a5459d2419d0e7b3ca1a136692c496a262eeec9523cb5b4fe9e82642db8bfc25e6aec13269984bd6590c00f53933ad84918aaf6661502951552d3d6d22ad8bd048618ae3f3ddbf4585c44bf59f9ff1361fcd87798a40d06adc35111274ff069c753a3df8c7f9eb2f8da707466a99fd43548a0eb1679fdb31d8731bfbc957f8f45066a58e8dd13f6bf91c90b7e621c246ca3d22c56761defa5f93380aa217467d8845128c6518a62c7ade8308615a9a8c39e79d41bc90ae3e33aab332e83be003102700000000000075c1cdf94d09f4c95969f2b57b0749f110d651c046b02e3590f7f8b21fd9fb16fd601cadb7fe90411ebe1c31f4fe78a43b557fabceac64b49d79fcc53937d3196b95a665e6f047ac98a4e34b1a06d23134d6d24f8938bfb6cf6d8dd2bde7ef7ab916b3ee31b95a320d4808b143370bf1bcf4331620af3b31a2bb44cdb1f637472d8fb7679e2a5bc550baa09866db16b3c63423ece4b18d152b7cfab3b0f6851f12bd82566168136c472d5420711398082a89866c986e35f57a95e3347b6a00cffd9802e7eec9b0c07d95f31c4dc377104f08b6eca34b781d6a797c538ed98d360eff9acc7534fff9c68002abce1c55a7349d39f5bb656e61f724795885173258d494207998700e7ac49852cfb29510d89c5971ae07ac1a040e3bdc765ecfb7fc2b431919367498a5e266024de0553f74d0a97799c0c09352ecc6fda5123b776974f3121cc6ae07926fb3ede54dd699996c7610061f25efd8faec938e43291825c92732802ac224f04988383fd6a063c5c76e8d78f1cf4a954e7ee2714248220e214294f8d334dd1944eededd92a69a1e03b109e8cd0722e626a23a843babb193518db895b0cd8402e9c67f4404539a46d86dedbbe30a70960653f0c0eb7c0b089c67238f4280562a216687122b366e16517c25c5b5c584b58048467a128e92b9eb77330af7fa2c6d2919640529d9ab5b292f98c3219cf035eccdeaa4d5dddc08cbbb1d82896127f38609e879a2a5ad4b11934877fffede32821112460fe736bb6ae19bf446ca380f3207c3f6bdc0aa4f0cd7cbf160f3a3a888088fe9991cdb851bfb08034a6bc1da7751a2c98062334b6e792f4853f7181af5e491c29cd530e9d5e18664d02e490bbd0f1f035c3587112a7c1858abc4bb3a0d7d197d546ca466e2fd34a05e678a9f0c20f470114d71217bf4be10577d2fcc47342c20db770331d31237b5932081171ff60af6b3162e86b08b988e1d3a801afe005cec926e18a75e0ba8e61d3628f6ac65e2f01020ccbd96cab423ffd4f024a96811b5db212887ada89568b06e6e1e779fc2c5151f303d05db5e7b112fb32324e10fadc4cd14bd1a2ab5f5bc5c0a0962d2af664ffe0734932d7f73569eb33fb0dcbe92b500e465d9fba6c61262737316a1294eaab03183d393ceb9d1ce36096af6e91d1b952bbeb3cab2e903fb02582229c3ee5161c1ac47f8d8cd6fc96959d889bde73fca689753840e63c1e2b8d69241897e85bc25d22fe5d874d4a2d39a2ba1a9e7327eb79f35468c01d91648fe34dabe117f5902318026e10840be18ce5b763edcff45af342736abca04d2a322764d94b890e115cafbae8ca4f62016b5956d19d20ea05733662f0710310ba0ac22c49211980c38ab9dbf39a2e8e4567470470d7cb4536fa90067bbeee16bc877f5da6943c2a6083f0a4f291d4642d89659c7027a47baf54f79b6a24650f709155150916db49bd4b7b9a084cce9cab997e197b2ebd82f30e3a29a9ccc067b9a8052a8db1a1ec70f1c76625140600370aee3d480c71d816588940e209d1829b19646c1b676f27cdf96bbff8836292e29e023cff2922ba55cdb9f973228cd12cf7ade0f671c6eff71b83293bb13a628a4cdc99a9bab716e8eb90a7a05ddc5830db7d477f6cc7a0dbf91105c9557a30225dcb744b3cf638a7ac469060356f239a7258859b9d73964e4d5cfaada1872c1f95d111c4655fe8837d34e1625ea6c6dafbb2a92f1d3a71ae5c5f0a13dae3028a0ebfd9cbfe078542b85567f9eafba3f5d2c547a3120bccbfcaa5ec2457ffcd3897197b25f0a69063a6968390d1b1ed7660e739cfe99f351a562e35c35e385cecd70e6617c21be09d0e6c8bd2f02dc7b32ea70ef7613b93c47180145802cf2c570d4702d03ef8a57ec4325ce95d0401e2a1d567621977a888d10917ee9a706bd73cea8c90f9822e21fe92b9fa37b00917285c0a0f35d5094335a1a951d204a815de677fbcdd9c90c5100a46a197a673e1cc7a8b702ebc0ef465616ac42f90111d6e50bd0922191a9ca956c508657c1b8f34409defb008ee1266a5e05db79dc1cbe5088cd38c3d99be3957c8592e0d16db7ef5779932bc9ff1180c7e26a429226bc5da9fc6b624f8550e5d8caa5952d1c6b3fbb5654e3952a1e975a769f018cb1be7f899708963e2d22850b0423131d655c9c560e6d76a75d1e2716c3f0b3b80e81174c6efba4f28f58581c1fa586c4ca4e85e981b691c60c71f6221c66b6c0b9a7b79a94cbc9619eeda90a8e8f6e254b45ffa2ef1e74911ea0cfdbf6ed1e8b7bb8be450fe509fcb65eb0a49a83c04754657bd1b82c58647b3ae5edcb33e711682adc10cc17377c071ff89c549d4fce8b81e1897d9d03ef3ad67eb2283dae232c5060926a22cef1540dc6658d1aaaa5972411ce0e8124c092213b14a8a5490ecd5379238c7a7cb64af48596b89ea0ec5e3ecd359bc035dd73d18de1e63c2b4cf0f0cb6adf5065cd450ade4218171468a1212c97342a0833b7f0d34c6d863c51f4bbd0468d607cfa96c77f10db2596e2b6e21287f4d07269928daa5f3ed2e0dd61d3d5b45c479d73b1df18f4e72060caf1ecd1ec9b3f9c421b6f2e5c4201c06eb945b2a6bb967453eacf5bcedb052bac375902598c7d68a7eb8499a46bb9b433155c36c5d3a0bfcb9b0079e81c156f5f89051b918330d0ad72a5d5d60b4b32ab45a3f72805199eec78be60d6d22652289efb89788af12b65f37ce78f79386aaf15040ea77ecbbe939f09a706e60b075987c0ee593f4a10c54073fcc23c686688b9b39b140456b05b68e9e7b183287e9a607a65615da2e377ae36da4b299e0d635bc5af350c3cee5ae7e52fdc691219d67ebd7772e84d46048b8256c7f292a3ad65a0b7fdd1cf94fe3e1064985f0abb7577bdf5cc19f30ffac488ef6ab6ee45e72b6ee5b3bddc3b28ea9ef2e5711ca48f07dd70fb7818b68cb5af2ca12ea93e6bc1711ca817e39172246721ae4513b6a05f1c6a41212b691d94c57dd4df8453edfaa40c95c9cd344882703181b30ce7ce80da9f33e6fc11c7aebf256deab7f0489a23abc82924716b4a657495a017ad0cd26f143279d370e530c679ee9240e307a9c0c1092d8bb487f941fe8cef23269ddb5940e29336c0f7dc59236fc1671f89aa80d8221536bf063aaa9f450a2c354c576047ed5cf2ea82955a628f9ba46d31670b2b5c0b09b0475d7cdec6b63b1c3c0b5d7e2a2bd5a0985bfa43f4093c7f584042496e6c6828f003530c8e563cf2cbfa452270fbe49fab2514bfc202f7d5fd1913e5eb5fc628a0011b05898f39a04d721470fd61eb1f4d6b75aab5f68f0fc036d54cc0d1a009160bd0d76f81197a5ea7b9cfa8c0d7b5f46b4e3124a1f801ea759bece7426a9e2e19a2d52cd40e25c48c334e5c8fe4cdf2585add0e1b6f18d84178f75c2f48f5f3ec757567902093dd4b4dfa37d1eee4420804c2315ecc918f136b4cafec0a0b7ad27756b37f0e8af008fff267fbc4b0fe8fae879e7b2874a35a04a803f75ff826169887572305e79d1c2314286b5fef1a39691e5f15fc0c3dbba95eac9351b831a2bee0fb690a018c5e16e25943731781e146845490202ce6ce096f973162b2fe7f55e9d6060649c56cd45175a3f11cf8d7d9852179279062eb996c413920f599ab88fa7f9b29643f4684e1b07c947a0e32cc677455dbdafd8f06cba671a27cb7cfe68fe0d91819a03c9b369c5284a14cc195edddc5b6e856a4a51571af70584ad4ed2d740a20b4c007106574e0ceb14db3c3b784d951d3cb6b8398c61d24c0b938f19b6b0028c4449d053f512d3eeb0788f90d6a93cfd64ebbec246044a8b79124160bc698150d4437939e1f8cb2df7496da892e6519a8b1b06843557c44980ea78f39dfbe237a876c299df9a29240225042cdbd4e7f3efc6592c8377b95a8414dafe9ce0d073889b3aa5c281f8a1f9578359bf8ba333f6e409d6dafc66a49692c4547d18e1c2eae3d484ec56ee175ca07518eec6a764a900999727e8620972ed281fbaf051a0bd34794a0919eb2a99a697c35b255df28fc30f62eebfe50123c50b3c740943ebefe14535513f559e479b033920368f0e4762a925715262abd3e64c0132838290463b07ac0df264a3b38fee1b64d9fd904d41322c4b95688f6dacea095eec0246eb0ef71f5db9d9a85d0dc8db980275fe15ccf3b33a19e8f55f4e280d976403455690ab689182caa94d566ed5afbcdb2b19bf029270443a41a68d66407d0850608da61bcb76809a2d67cb410271358f3716fe67097df39f3f6843ae47ed553219acb974976a50a0d27be93c90b5b48166d559ab10287e40c5868d610bfc73d170d2c569df5ab95b5072a91b46eb87eeee9863566f416ba56dbb600c81de7b80ff86c45e9f4929ef78e6819a60ef4c7119dad4bef7a8f579304bbe00f60d441147f77c713599d31e96cd597fc989733dbcd22c8fd4b5507a714751d917edae41ceeba72bbff4be62f1d536a0c3b686016a5868bdf71cd3a6d3c57068f6cef381cbf76363ba9cabaab0b6268cb7d465a9348c5361f115f7c0ae9b8923f7858a92e82960b4613579c16ffc809c32f14f128dafad5f088b99a0c6452af397be5b20eb211e0ec96d745fae33db6cfe13ba5f414c67d3fd636afe327660f74f59c2c2a740d98192e18bbf1ec195a30c687c5675283886d826a347c4a9ad102ba255910f41030b9eadaa55ac5ea0f2a66c571bfbd37535840e151b03464fbda28513338b1b70eb78c16181b98c5c5daf757fcf2516a65f48f788aa6d3ea500c5ed12b08caad92b43735a50ef442558a7d1a61e858543f385b3deae082089e1384b6970a9f9b8bb1fa488f522a8df4626889498f17b2cceee8dcb428e5e90f8070dac323bd6f839917f655083939a23f559dad98bf702832c46c1911a360b39fc6d5110f55f60333f28529c2ad3241c81386d43b991d0fc5eb75e4558542356999350e2cba6178254bcf439c0e4e0171eea8bfed049099b9ee5e2bfd7f3634e31c6a352ebf1fbb4936aff521010e83c7a4b3889d4e6e7bb77234a16c90eecbefa8746439dc15f1d420a716fef9bcfae21a183812309d3a1cf7a222fc4f8914078af9c0287b0c91d3247e1248d1f9bc4b8ec38032c88bed2348d9c19cd1ee03b9b429003a1d756daf9fc0f4f8a490f8f6f093525e8fed169a50e81dc38977ca7082e6a81159d43bc1df5c075e092d18dcf9376ff945439c0f120a5714d397a33877c16d3cb6340eca31660e7a53843e130c5a3d6006d10dd7dfb0d8f72b22e35085d63015576176c1511efa6ed574b764123e257782f6607968fa817867945becb0671b09e810bb24c78e4662d28297c645012f015276fdd547ef6e201393ceff4608d10668ae3046149ee773e472696288fb6c33e97fb571410fb735838625a6118e7534e96bdf2a26964bc5b33ef9698bfb2b51dce97a72060bfb4cae9a55811a05842f27a503f0dc9ba4c1ee1085fb8157c61bf699c4561dcd0549f98fca88002cb909915065d78d1df7f70bd12d7933ff17b572620453815f200e283db2a52898661150aab0be7bccb1833ad0f798edd1baa28806ec86b0251e87682fd01110735e0471103c6b0102afc584379c8d02c33fff8037b827d542e6f7321b82db4631d81e02268056d89fa046a757d4073c9021071cecc4f945adbe0043289ca41924f0225378383402e6c3439c9d9090b9e6a712df454d25a2fe958c954ef4dd2e103b1812d51464ba1d8ee8cbf1414a5dc7f0419109705c6d837b919f118572a389543afb418ad8253431850b04badc4b9112f7bef12adcd0465f871e80cd9de361831cbffdb5f7904cdfadc5f12372a9f4e23cb642c04a17d1194d565b784e01342e01465c5a086c706529b89a8ab23373a2acf935e2a25e2dda7e61314b04090e6a09e30af4336a162d447520a190372dd5a4719ed7b3df35eb51fe9b4a4449ac5c01c8e7f8b4c56c372f19a00b69924704fe6b22dcd7d02ed9cfbf37452a0481723d69ab7803e8827a3dd159267e809d63c746dc8bca575430509b93661b0224cc14dd349ebb5181436bc5991a2921a61a478bcc2efa68194b3fbb1f0159c5024a096c87b659a87faf4a719de6e66dfcd3eedf98765d5d30c2d5c300fba7c4e40d2aca825b3201b84a2afad5a22cdec3e7b54fd3d81e94ff23751fc239b0ff697c1e1a801e38cce6519107f27c461e7c18d06c0471fcbf6c4a279ea3a8b52217a3147ebd2b43f568907b7978d63276753f80e4b22be152a53506d6fb772fc3bd6300c6420e034d365bc70be2b2d4d84ee5516d72056d112b675d705c75d21c5fbf3a124bc833a90472bf366c2fcb6a6b9bdf464fa723dfa4045cb846d256f2f3b438bf0b7fa9a0fde88f519f4c95d212ccced342a7cbc4bc4438e5e06db92320e3047697ab0df3f3166ee7c22addf783d5161a6a656e8e7b8c8bf6b00d7d5172ac1908097bd1aeb7dcf325298057f7f80ed405a5c27a81d2c1578505a430393f111d85e6a220a3739afcda4bfd56962251eb1e76be1de3fdf72fd820f126d1653534b5226295eed8dcc24be4f788b555e3fec587d3fe0e414a3ac7bc6f3a5ebf480eb441dd0ff285dacb6898f2eb716360b2a8309a78070569ad34643f25565c342273d1494cbc2c2f9ffc8f2ee5059b9a0b0f0808a0f22d94d729b68996c9df662d47ff5ba58fc330038e83a026af33f3ad731cdc8c11c0be16acec8b80ca196c300a5d3fa807d90ba958fe1afe2041d5320bf153ca9b3cd2e38a1d443bceed2e2b429a829795d03f5452b9db1b9988edf983054e1988cf7b4dca49f6ece656cf19e60fcf77458b6ee839ebddd4d347d006ecd01167d502f692090b96c1947cfb02bd3b85f13fe1c32d7276b797cee4628c18ec3af7c90a62e32f25f879ea953622ec5955e1f78b9c37e20bd55164138371ab893986e93f35a467c2512bc09f4406283ec22e7e46f51bcb623dd2c2c626f3b06e6186a5132daf5dcec9c231875d0df295e9097819f0fc65396d566a70272f1648c0e494b0f763140038796b4efa375b9a1135617af1f70727149c109a84764c2bf3d6dfada98731b7c293cbc9b91510f888e84c795df89fa4d93dd2d629ff3e37816861e2c4906a19371a9c29f012d7982b02260d58e8b862f9a6dbe2d398366f2e81a39fa6081e2caeec42f2463696dca820f1d423e6ce56225bbda194acfde4b1084586baa85704488eaf8a973b1edb3f8907b40255c6acea53938476535b326675674f05af31f5682181c56411af346de0104338b4659e4f5762cf5d78d858c813d0e750627e4fea737f91d6009c07996bfae7baee96e24d2541ca4db94186ba7f2e3be75ec2bde8ac9a97b33fce849d1ed3db2962939ebf7cbf15c42998ec73662ab081926bcb669f5781f73093b57575cd5e445b10f1aebb9ec4294fb90dcd1fa9bddcb20458c704a72d102c7d2438f2d14ec69309c91610001c5ec9eafaca6fc9f7e7b58327204d15f3d227c9e1e8580a08c80703eae70d3e60ed27869175baa1c0d4ac28faf887df32c7221379232e7ef4197c90ab17815a9bb638ca121e14ebcc6aa5720cfdd5c4311922efaefb79cf15a50a4742c53c3465e9ad80707ebca3b62b26b51171bf60403f11740fe311b1aa4c2b3f11491e81c5e056154a59b54549daba94aa4a29bcfeb6054ed06e18da28ad2d4f7604a73c6b0a51a67d5e6cf891a85e9f18445f562abe24a248795a1679382e44f425d7884f8e8c13827d115c5ba65cde5885f167d05a132a961d5965eca90ad9db04d0142e69d4cae60e5b460ab2efd81a864973db3827c7d4242a593a1a11321251090ddf005dca488472d1b353f1ee493202cf633f0c9fc170fe247c3f044646b579dc5575e2e0b7810c8e26172c3a38773feed5161364c672f6217e570f322b907b262821c5a1cf8afb39e2bf1caf1352deecbb421b78a50e449a5dbec4e30427b4e8bff73fd4d5b2db374d77129de2de3c9d6997035cedad75939b46d82b013774b5cff5199cb8d59824edb940feebd588446b8b24a058d6ec295d0990623256831a39922f0a859d11347461e5580a60260404fb33cacf86087bbad6af003bd5c12cb2f75ff9f7230ae8fb47fa69e6110ea1716519419bd5732aabca5001f4493653d11c585c67b2e2079fa24363821cd1fefa1626e5f8b44f27f1bb39e741c766a1cd1aa104c4cc23d6ee66f451ff1eeca5cea821e6b4c1bd7ad6c5ada4929c983c3986078fc535cd80b5d5cd8a126eee5d5654370e80be18d70a62d7fced2b9dc6462d258f83f6aafe4b566fdbba899cc1a5a03c8af871accc56f8bdc00666cfb242a7d571f882e68ae13e634935664946c081284b990a43b04d37e98b0a03a12b812381ca02c0b845c35c602380886bc9bf093341b8b7ad89b9065d1589e1171cf33ac308b9d7085aa8c0aa4a0af1693cc8f1350913f714759c090bac01225230f9792f7b18f416c77346202ae7f740c936e03d96d4d33783289c0e638c63c480c8049fd95bdbdf3f97afab8f359f253eea7814300bbfd81ac6685be9343ecdcbac0c8669001f0556e8caecf21fa71d03b857030d63d8388ef9cfc6f8bf45a66c8979ffe6b6f70d268cde3062fee59317985410003b810e5a5556533b50304e5bc12d4782cbe430e7061ac47f4c401670d7be00eb7cc8c636f81d7cdbc92290fe686ec2ddac04f7860b9f36afb6a69ac3e535271a8bdc5f23fcc77b64f194520e0f98b3e21b4447caa2ea6db76ba9e7646e5c25031e71510f021e8552bf52dbbe23bfa8035ac38d77d8738f0b6cf65771c48d1912b4a7fbebded0d9c65d7d47e7fa54bff45d6caceacb7f1eaab0850d63d53b2e74319648122f13a83672b0b0014973a30f2095cdf29cb15f354b6010266d2e37daf1cb0a1e84ff95b0ad3f7f20f103d8563429f89e8b6aaa31fbb65bf7ac92ac2d35fc973a719b6a3c0c897cb8f250509c42e62ff8fd00e7755fc553b06c1a33ce3dd6ba0ee64f9258331fbc549527146f18a8e027d4b027a5c39e1558809129df97b09f74048cfb1a07c7ece4fb59dc72341bb157940a4f6d369abc7cc75a098fcaf8f034edf7043807617c3ef5d1a303cfb0280a319a077a3e6367d0838c2d078eef8a6d5f0e717372995004e7ec752c9f93fc0166154a76d63d8536867a333d6337decc12cc88e0d4049002fa16481a9e111a65800e4a0d4d7eb338237b13595914541694b9965336c6831901e08bba44e737e0c0c2fa2ae1692542111da47c0da4e6bdbf325180d555554ae789c35a45daa8c31526d114795c966a341c215086ddd507bfb06845ebc574bb246a0d455db2be6da0b7bf60b43af324ba1431993c95da505f0f228cd024154bf2ce3b24a413babe1b5000e46ee2aab9acb1bcc53fcfa05261d71ba0231d71a45f68a7b498ab25e498cf95af56c3aad94efa2518d5138a600b6d4db8c1e7c0881630b52ec57149193d3e5c1f13657e5b70119d1da5a8839f31566742c9cd40f46f44ea758a02beb0b157f97f323c40502703366be41a85be427229cc814959f84c129aaf006b5fa3d9b5d40cec932415141e864e31b2740c54039513fddc811dd4a77147650f08de1ae9e536adb0d5c02fe509c78596758954a06d9cbf4385599813bf73408c568822d1a4b91916ab81dcf390aa9fbadf3d56f1535d02443d6074a7beefe55edaa08c7743dbf9c7a66490dc2219f87b88c5e4a3798ddca1d89c4500de563ae4c2e6048397f9d8876f851e61a0a011bc776b329e31cb89b46aeb10d95567319bc0803f3461c57a235b60ae7ab7678f1c2b55a3565f9dff4ee709cbcb97495c7c1e817d8b170c9b0f81ee6ab12379d4539dd17d7551904bb54bec40441f573d0b7fccdca220822dc80c911a7fb045447a8d2b211d396b76553000a43a4a3d002703186d9893b3ec94e6c663ff0acc7f9fdb27220610cc8de4fdb0f80b68837141d112be981dbe4b1d4bd6008cb974711187af771fe2e7aaa7624eb4d249ba15d1f89380a7112a63e7413cc183a2a0e2dcbdcff5476b5ec0f62a484f611100f3cdbbda6c7d648a8ddda66eff173110b6e8b13f3b03a4acaa77bc4399a1f1947d4bc94500bfb01ffc87279a6ff6b11a95ef6be935293263d3de97750163ec95b2a5300f240c997839789467076b8d480f7cadc9e6944bd4dc37c4e5520e891b48394231799501fa5c1dfd0437cd369d49efb0118e8a3c8e2caa9303d3c70380a85631cad6b543081f00809d57931d40da85451753c8f6c8d44e1a1bfe0cdba1e67e028dbc8b89b9a61082e05eb93793d98b8a28dfb15b15357ce13fd0b25306455e075bbb1a52411eaa28ab3dad250568bb884ac54be7f3e3840629d18aed139cbf2b33ba31476fc3267d946e2832cb3ea53689e4834f3a17f5b90ebc59429223674f4bea263b8e922a5d37da8414d3adc1380d47c62258b5266dd472d6e95dd0999cc41dc5b60be16b800a9ffb0c6426739a7a7f8041426556860250b0f223f89a0033dd353bc9d7108842d9b3104ada1c9f0699adad68ee478f11716b3c9ba69dec0fa333002f8dee10c8901609",
-    orchard: {
-      actions: [
-        {
-          cv: "28192518b0209ec32d9d83cccecbcf5428c156c774f54878dca774b2fd0abf99",
-          nullifier: "bb2202837283f8b6346ce90ba21d9c9f3b3396e20643385fbe2302326b548708",
-          rk: "6d34f2c81cabd778580abe233e2b367a79fe2b23b1e127508a9a6fd23676ba8f",
-          cmx: "fc2decae6f69f41d0df41d5a6db4f6c915f46962e5d4209b8ffe8df3e4fff226",
-          ephemeralKey: "3d350536f4f3d826ed0bfcedccee16d737df585a56938b65e3587c32c486bdbf",
-          encCiphertext:
-            "8dfb2a622832a346b79fde21144afe6436e1ea4e6520dc4f9988afa57a910be27f4614b3e6ddecba2794659fcfb89d9d976253d5e555b8c6db28d5ea833bf6f68e114bd80fa0eb14775f660e49693eb860659d97b1cae54e29f6598e6e40530a523d8c60e1f65dc265b96946abdd0d8844a5685f336d9a17c4fcb1f957697cc63b962ac27a655ce3532f62d9bdf17c2375d31f32aec69d71c8375ee9a1729ae3f081184cf55a639a659fbd5e26fbf9bda7f52ceb589bb39002b6fad43fad12d2a16ab0cdf34df6290bf69be7ac6d9512a181120f42d9fc7166f4880c3b2393a7de19f143a538bd03212e1f3bbe4dc2e143a166587f3b4fdbd4a0dab51cc71cf95d59866213f49703ca43db3fd69215ddf985c3bc31a08b4ab7b32f101b4dad87326839f148de0f59b3b9f820855ab649ceb0cd89cb27687a2991909fc5d2a4996ddad241e91c1fc53283c791c02a2d985f09ba47e4b6d46a5025b6fd0c335edab9a4f7562431d59c22368b46e20dfee70fed4e940f82e0eabbb5b21ccc005a3c3939cc55a3a362fd48efe5b8736fc2a6ea2ce942729a316eca35135b53ddb663b37d93ae64249a0890dca4ed86c72879cf248ba92430a8f3eb57611b7d4d5a89c9d517ea8cde7de7cd7f8316807298d2bcf732b0113a92912b87074371a1b4c353a996e15619bd813ab9ac41a695f1678c7cba0f8806fdf1d8c1d3e07d369f6036c9df71891a27f5f44f3a41b246a9ad5768cad155ace94ee732c2fc3d3162019f2d62295b2cfa4a62271446e9b7ab77af001dfc6b136e19dced51b5ff9e3f253877edb7",
-          spendAuthSig:
-            "93d98b8a28dfb15b15357ce13fd0b25306455e075bbb1a52411eaa28ab3dad250568bb884ac54be7f3e3840629d18aed139cbf2b33ba31476fc3267d946e2832",
-          outCiphertext:
-            "e62b1a887ba55d7f1058d359929b93963fe232afdf56b92f1ba6d7d2a92300f4b45bf7c69366bd708edb56db4192748f368a05e76c002e9c3bd03ee91f92610a9310c311472a922a6ed4cfef760277c7",
-        },
-        {
-          cv: "bc162c43f73d6ca8ee4b8748e8e3473b237d15f78f85cd8969289a18f85a0f23",
-          nullifier: "be3f5f0eba8529f434a54e363ad7b20901dab8c7d92918122c1d9d066fcc9706",
-          rk: "38527f3c5c67257910e55f30de035941712c10957afe2bec57c0644ee59bf930",
-          cmx: "15bcff740c52d7d6647fe81cd8b332e5104b8bd0b0ce685a58004ad5b8279e2c",
-          ephemeralKey: "53e654a951c5d7eb0a31fcb1e554907a435d00f6473d361cb076a9229ea41626",
-          encCiphertext:
-            "5e0a04002ee79238fb32d00529dc5198cf23cc9456023861247d8259369457e546697951141f3a3db58f327b191a765aeb5f8cbcb7e6b2024896a4ba3dbb462e008fe8bf320d52afd8cda355fe9a42ac9d3b11351022659b277ba59d204bd10d5b36532e8fa11a98d7ee6a9293b033417105be66e742785f3528343d987c33ce2e5c9dee2d6c49873ecbf8c54c10fd4cf27721004eae3cdbc61854a78aec2a1dbc552989dbf7dfeea583fa57860083f2611742c09114c7ceac786f4e4e25ddb4a6ce83f431efee6d1f8a3e6675e7185431b84056740ceb07b5ae2277d7efcb45739fc5cc90220ac59186294e466a1bf37b231e7965699cdbad494db0ffc3fc5c2518d412abdb6a80cc69b64504e1f29ada9dfc5fffd89a8ed524fa059bb8be6b12b0636371ff5dea026e7a17c1f1595d2eef4f876bfa33e63ffbe3f2677406121a661c0fc7c4f0bcba7a808f5d8c146bdccf436988ea521df47978b6db1dfcec67a5fc8b71ffafb97c8918d640ff37a66ac6763394481f5ac79c4a51a1a1038672b0e69b675b0503132689b57bb57e0891cbe25aacd6e6ada436fcc3da1003405f052571ea9d5da929a89e7b7f826b4ff622477a2abe897ba8622d5b4d221a6b2717058bf44fbbf04151ed9417fb5fcb2d381bb0b6a5459d2419d0e7b3ca1a136692c496a262eeec9523cb5b4fe9e82642db8bfc25e6aec13269984bd6590c00f53933ad84918aaf6661502951552d3d6d22ad8bd048618ae3f3ddbf4585c44bf59f9ff1361fcd87798a40d06adc35111274ff069c753a3df8c7f9eb2f8da707466a99fd",
-          spendAuthSig:
-            "cb3ea53689e4834f3a17f5b90ebc59429223674f4bea263b8e922a5d37da8414d3adc1380d47c62258b5266dd472d6e95dd0999cc41dc5b60be16b800a9ffb0c",
-          outCiphertext:
-            "43548a0eb1679fdb31d8731bfbc957f8f45066a58e8dd13f6bf91c90b7e621c246ca3d22c56761defa5f93380aa217467d8845128c6518a62c7ade8308615a9a8c39e79d41bc90ae3e33aab332e83be0",
-        },
-      ],
-      valueBalance: 0.0001,
-      valueBalanceZat: 10000,
-      flags: {
-        enableOutputs: true,
-        enableSpends: true,
-      },
-      anchor: "75c1cdf94d09f4c95969f2b57b0749f110d651c046b02e3590f7f8b21fd9fb16",
-      proof: "...",
-      bindingSig: "...",
-    },
-    blockhash: "005290002b3198cd350add1abed3d0c53f736f5ed54e3ec463c5d017e602630e",
-    height: 3795067,
-    time: 1768842685,
-  },
-];
-
-export const decryptedOrchardData = {
-  orchard_outputs: [
-    {
-      amount: new BigNumber(30000000), // zatoshis
-      memo: "Thanks for using testnet.ZecFaucet.com",
-      transfer_type: "incoming",
-    },
-  ],
-  sapling_outputs: [],
-};
-
-export const decryptedSaplingData = {
-  orchard_outputs: [
-    {
-      amount: new BigNumber(14490000), // zatoshis
-      memo: "",
-      transfer_type: "internal",
-    },
-    {
-      amount: new BigNumber(14490000), // zatoshis
-      memo: "",
-      transfer_type: "internal",
-    },
-  ],
-  sapling_outputs: [],
-};
-
-export const txShieldedSapling = {
-  hex: "050000800a27a726f04dec4d00000000a6293a00000000021690546cd26e6db813142ebe6f85cf1d526693f31ae575b473ab8732558911d8115568656e8a9e7a608bc7baf845114e1cb46d25fdb675ba022ff426cac5e35da02d26d53dff40661552ca76cc444af4b5ca121fc364c8126e4cd8039e3021975a300d5e9fc9661698bd149ec49a077ce9643d0deb792e471799d94d4530a8bb01bcb988dc2ddd8fb5030e44b4b6533c965b7c82516e3539da5578e27ff2fc30c47ce646584653f0fdf8743bbc812c1fc32aa6c2a2a04111c63067f51d3032befa669de4546986f70addf13f4faf31dc52a03aeffd45d8519dee687f68ae904e59fdc1c20d76f2267b82506d08810eda92993284fc5063a40818b83fdfcaa8c24a8f735698a384eaaa49bcf84d2868ebd47247916b00c77641b01019a3181d5c287bbd6c96756337bcd15a0717d7ce84346bfd30e32d5269bc79b664ff6de072e2259c283b13558fc568f065b4bdaeb69fbf4a85898b603263d6eb95914d72b224c72dd7bb0ac22282ac9b60b647351c973319a4ceb0ce9389a83ca7efc06e31a6ab0c76de6be4aa0eb786be646bc094ed2ac7f6901e07035c60d491163589713b69ffac244d2b50f2e8dbe14c8b6f77a1b1453aafa2b6182229084dc6227824ea19afe71677aa4243cc5fd3b1de6b190552f7db39a20eee4ccfcf02e6aa50d129cef00af8559332a61847ada5f6d0568348283aeb76fc5e5844ef56b0c2b8392b3e9f46b9db0fb1e819fb6b31c2776060056dd56d268e96b02eb819459ff6791ec3a6b6a85ae7aae1bb252d3c32c83b22adf2e5b64325654c5e1b67f78da4cb4a8f52e7735851806c15639a22329d6bc8fd9cd1b8545ceb397456299050daf65729a81a66a2ac4a353d690cd43d79b0d032e3f801590e3504aba3d9c6a751fed42970ab3217f0b165f3bfdfa7d4aa910bee648457b0e100e0397d2e4890cc16a54ad9801feba09949788b797967cdc725dea8bf1f3b6feb1d756b7f8c25af9e78037460ee1158b016b0b83c1786c51f2e466488e96db93850f4c77b2df448f74ddc2cefe4ae9eb50e46e9d6eadadce9eccff1ca6bacb197b6cbcffceab1bdb51eaa8976f28f5870feba9e632566981b0eae2cc305d8c5f20063ee98c2cb99ac5caf0b69342c59af863f8ca4a956dcfd5acce26cf268deed0ee477380958929347fede2355aa50425afd24f2f224869cd09cb01731b0828e8c8cee0dfb039bbcf2ef0d9fdbbbf2c28afd6757ce2a8ff2986a4a40a6bfae7140146d75e1cc34d12af10bbdcad93b7fc71cb4d01c5391f1f08e0d44a99f23d93363fbcaf44f79fd3ac29696c8b7464ddee5d18e839d5cc39ffb53f00dfb9fb5e366066bc0dc0ace0d866b6596aa21108633c31ab7409365dcb13b32a446d5e159aa6f6e78645e63763935221ead9a53abdba851f18545c73a9b62c39931747910488eb9d98007b4316322addf4a78b1cadf1d34c53ac9e95f5cff1cd5c32ba39884703ed6330a0f38487a0a4fe1bde01220e29a4d3fe17702810e8dd065c72f44c605d3fed4d01c983e7750c5183c1610d98d1db700d4d054540288a2233c2385086f9856e57f16f9251a52c60dc45c8a8897ca243c2d796a411828b7efa98067e41a59b78e2a1939ef03bf820048f60345845055d35b379be86091342ff909bb7cef5eb975ebac7dc3f6cc88696871710ec870fad29278e0ecc28e1dc5cde442815f048b48cfa471d88bc8b6b12f2d5d45ccff193b8e6c52fe23e2af918bc16cdadbdfe6a95d583a3e45433c3e04bfc2129abc5786753f7d9a4113b379e2d0222d7faeb4cc2242316d31c91514622834833407c40637f47b1ceadbac4f8252d7cb179c4ff25f80c6a4a7647539e22def9eb0c25f4fb806e099d1efa4f181d9afc367c6d438a6e5fc0ab3ec6d067233d4e4fe1e4310dcfa7242627be3e173e6386ff67b13fe80c1970bb8a41563e0fd081b401ea6b858ea579d47deee075ccab4a54248be2cf645f4e50cd4095dce2268e3c8a4ddb929db526fff7960d9a8a1d009adaa7d68111a269a75b59eba9f1cfacfc7f187324831a39aacebb83c53dde42cb5cf00ade06db91dc2b2df29386cc1e4ce77ac33fdce9ea6a2107f077087293962790853bfe7060185c535753503c0bdf0ffffffffffb3da3e64a0879845e7d039992a034713d159c95d3d89d815ab50f3390ddcedef7271a5e6a611618ada58f0c028c44a41a3ace5e8229a13c2d4f7af1d72202d73d1bdb4d0c05cff5a5be83b922571d1d10354e09a7b03483f963fef4f6e940e711177933c948009ef62b53622530717baba8a4490ad01121df87eb1adee592978d645b418bb85eda2c934685fa000b69db25fec9c3d2eb0f65d878237814d726215d8bdccad1d7f697a8005db30000a43ddd2dc483c1c66a419dc9e154b8430689187e6ebc1a93dfcd4422b1df40a41ba5b86c957ecd9966d6915aa1df00e2debd2fc4fea12880ae30ad4450cb8cc23529406f5de55f3be3eae838a1c5b7d39d1b8490d8c51868f73e793da8706027c699e53c9a4d5b48fb1b4a42d3e170e90d00a8908f9af7037f1cc28abee6ce965c284516c8111857f4eb6a5f39512b7736ff9265baabc9c03720172c73553e515dfb85bb26179e648da1606aa1663c82fcc4eaadd5fac192d7a41641e3751097ab42434f7cb4bc974c290a14b1e48df5bc566e4f1940964972c3227fa5e103962a220442a813c5f30f063337de014838cea34d2ca3a34d8558321ffcb9f910ff03dec73ce9192340ac4d7244f731fa54a0e02277732c68c1284fc18184ff8e19314fa9a76d81fee29cd5bac23257c7a47383a1f4f484c4f08999caae074fe19d769e8129dfd84b2d30030f7dfa1f8e8700d38b25f70393160e3e6447da0877c253cbe74c9e36176469fe9d701054d30242b38f6193e0a8079d171c5b870651b258aec6b2afc8bc35add4d8e75f94b9caa292c818971bf188dd7e98a1008c922b7327aa8e59b239cecb02c77ffc569e3591081e26a5ed0f2068b14e437b71f2075d4a51f92f5a60e96af5d965636aaaea5829e25b65bc2592445243d0f160792971ff27b7f517d4892153b8e95c345e5edb7543d2f8cfa50752db0dd114e82f50182ad2a10789aeaca20eb383b17b12ab431bde527de167bb1fb097b2f84a57fdf10c0423c80513059c85330d6159e78a46a8451ac731e2f26fd2e217d00e5c13e9f062386b192318942ee1f9181c4f4e75d3ef70c638e2c98a1088ce58e4d8fc22cf070322921d2ac06fdf24250c2ae28be258434b6673a33cf1d87cf65f4a4085d101a4ef2db7824d1e07982a562075a2081dc3cc2b70e66c444d581407722ac96301ed0993165a574e13655a9383c0549e0ad2b202194916f11152a02f65d487e1b6662b9ed9637c83f07f46004c8f9d853d444532c4dac5ccb193ab98a4fcdb0cda57f2ec79019a4d09a3a31088451c6de1026a0f94e1095e3b7db408a39bd67f345099a1c63d020f2aa81d5b05e02b436c2e24d74aa8f84b9f3b29f204c45314b1f65f57135c9ceb82991578c9ff9a076ae55af6e051b55b823720aa71ca08a7b13444335eb1dcc51c391c99bc5347412327550e998836e6534eb4314a7da55678d949a04c7dd2489385192a2d1976ae935193faa402e852eab40c775672704b66c389b6b911be95102e0c4d8ce988afe85663bac10aaca79a2bc52babf52f86f1de6ec6afa3b0cef14eb8031837e32ede476573480d9b1427eda50b8b4994f280acf40f36ef348894bded313a83d6dbb624f263deb35cebd8a7c2ccc07a4be5b5398eee2cd2eff80af3d59534503519a34bf2aec4f453be65fe3459ff50d8d025c378d55e08894e6f4672132a5061c748e9e61edf1254c4c51f9a37f37a613dba65d1302fd8ae8faefd8f21605b93903c29919f1d91e3518e51cff3dd40f685d941a3666a34d416014798ff98854c2fe7d0348f031bd079d3704a7bd098fa2b6bb1fbc0d6dc0bee6349afd2bb16c69d2889a8fd1984ac2765ac85990ce3b9c2743bf4c3a20cbff29b288409e287983e1dad257f33aa11093af0d0d52d8ee69659407c38780b2a70613fc94dd762b160e152b592c570470e9da118a9aa78ec8e18ab7263b332b42c1d63c525b48eeaf7c9eb330be21ba9938560af50e1a0d01d3f08df33092e3fce211277b6222e1b73d4f3178a6fb5a93e30fe30cba781a27f5eb71aa9328018432dbeb16fdaa61567bd3c2419e7f63852253077ab3fa30a8f474c503842997b4e84b26eba532b6afa16c665b170df166822a847d6b835bc518e268c7d7bf1e7c41e887df6dfae2d6382ca4c432c6faafb090ff79e517d53a2fb357d201978c8ddb745b483c4202f4d3b45a041c37484218fde27ae170ad2b1663a9007abade4f66d52fb873cd8e42756f2ace3b55ebc15912d462f85f56846cdf0ff1b729cf24eec702a69da93d1776fbd106d8fa32e6e627618840a629084faffc8fb241c17924f54ac682595f874f963c183ce5177cffa5280c17c942a63e8e1aa6f5058f082f9e268ac96c71d6b6eeb782bc1a4634468d5ce8eb8075ee71c1ca8a6dbdc903e47fc2aa3bf59dd9f1800cafefaf1a615a88014f61b2800c44119c1949515206b87b124e15fec2cd4244c2d4f14a9098dcda5c34bb328187d85cba59d0df8309a1b8c626da0329f55007efd560c08cd79720d3e59a5874d438c723cfc8f424ec05a2f4015f92ca26211ae44a581b45c766b0495400179ac357a1e44aa471cd06ee964833556d21e0d609a03b172a7bdea5d0683e982bcc60b683b62c44baf7997197964c2da2d1ee938dafe805770c8feb161e11ed97edb2960d2f191ddf7e88a7d492aebe677f62ee92eb10f1e0fb687d32fb6dd5aef5b6d138040ce2f0fe11ce23b31ca0da0950703ed2f020810cae9fba8dc02c00bb8ae936f4d82030bb7c9ab23da5c04b32d99697f53e7265c43e325b3cdf480c0915761056640afd3e7d9ab47c75664e8addfc590a7afe243f02c0b583d658259d6d5963198c30c5768e4f345c1bcc2c453d1dd0b1a6e964a7f53234d3d7e651dc32a5add95c04e81d7d249821962828545f15c614c14d92ddaa930360900f00000000002281e07dfc77d1d5c77803d3f4e3234113d3d7ec0625d942ddb5e0106d3b9a25fd601ca1119b9f0eee0a67ece53b362afbd3aba8cf944a0460fd33b9d32085220f1a2dfb6e4d0b03a8695c63e4b4f4608d54047edcddc1ac66cd0196b04acf4494833d664645ffd1cfd8dd401bf33ae769c66df17362989bee3f09b618f560fe83b935736ab3a1249145b2fe355ba67a137cd38f3214aba982a75de97746889d1ba4244b64589f03ae864c082657ce92cba4696a1210135b379c96f6fd2e60cac678ba2aa059a9044c29cd7623c04aa85b765919770f096ff7813976ac0556d099a18ddd95ccb15a86ff41965813d5701fbd507a4f598ecb6e6afadc9e5d6ed09e5e8d5551b9ba296145e50aa044dac915c694f9f75f310bcbbca96327a8f0db25e1b0001b0a0e13a7c71ccdd6d6ea443afcdfd5e7c75869b799e3cc781e6c106a87389f48baa90287d74c4dcfcc722c63d713bd466f80d30239fe5a7ded62aff07e1698c65abc3ee29c0b6c401317d50f685922cb16992d0c073f9b901a5000509b10f99d95ac21a57e099fc3a726d2916a0e4ec9b0948cb34b4b46607aa06e797387c82cfa37a2215dc354c517b87e7d7a83397c5ed7e203d9524556ca1813a9d719adf17c79a541b51be4403372e3072b496cb2c3c44a1b2e744ea7d60016672c8a858d23aa6212500f6812dc2b347bfac2335e6b91c05280023781aa12c80494871ab856a0da222f18fe9c64bec940ddeb0b9a81cc90079121714666c96a32181d3f5d87b3eda92edf810c6c98069d9fb54d1302cb9dfea191309f290b49e2c1284e47dfc53e0e33b16d84856bb78c428b4a751c4ae7dc670e141ae3761b0a56a3f8b16a7109a0694a3a71ef6bb9eb0895a81f0176d1c8f9526a40bef9f087838ffeda9774bf54cc3dd91eb5a4b0fafc13553d1deed89fec5661bbb5db6d61ee89f510c3090eb6a244de66381d60806d275c53c04ea9d5a783c287bc6d94415608fcd7fde2b0761aa383034f7364a7363f58e2b8a14c6da8eb037dc4c13bbcc112e66a2c56aa0d8bb5001f6aef0c6df5d91319e4de2b0cf009c7c8e8ae04b0502f9bb591336deb92df0fc34c59051562cfaa4588a7f2cb83656cf2964f2661b1947a48805f82d9d55a1fc88127ebca73fe304d91aa5d9366f466057534cac10402451d5d24b770dcdb1fc5bc03c6cb7d81b29bda41be4cbea154521ab8ee68ddb0f264c53de07763b5bb96f3cc7cb149715ce176f9f547ec9d6f0da76eedf818b7372740b24908af271dae5d27375484603e8ca5fe626ad83b31643d69aae43184019f41bc324d3d8618e2ffd2d4a7e5ec4e05f6ded7144f065151b619221ec5994c78d0777c62becc5001611614e4317c1adc2afdec0112595a2edc85ad221b1b8296dafbfdd1e7972d00e68e0fde50631c8b092fc347a1d0c4cba039ee20a29ce21584628d3a2946dc543b3b8d366115639e9cfc8ee61658b44effb8475779bd41f95b3d9514f68d8f432b6a43fcdd18e792ad2c21dd6a120fdcff97540cfbab98c7eea658ccf19f322ca22526715574e62c0efa92f4e888a9962b6f8e9f43b381cbcfa228124a2491aa8336b50b28c23e6e07515f719681d667eca0a45d65012705a1573cdf1adf5630f4b13cff62f97bd7fcfb8c41fd923736773679799707895e9c3ab494ec76c315a5abec5a5712c9d48d167d379f4095202bc0cdccfe04d83e12de29b7a9bfd41c9059e0a1003f95461a748d3b5f06529568429875e2b3f6c11d9c768ca4347d946ce3220269f5840bc1c15e39df9b57276d2762eb19233da4a7b6b4de52b09c9f11a2147b652a66b4e31fcf9bc6fd85ed78340857302d880391224c7ea17e6382cc3615fd61c1198be5296dcef5285b6d4990051df918ef86569ced92d1259e6e6342d7de5e6bcaa0b4500c9a6952245f0de11f697289e41eafffa68392dba489e0ce76389c66b347ae3ba340a52dbe74f62da5bd41bd7d5fdbe1f855483b31620afcbbecf85e1dfed50b94929b93b602a0ee86dfc021ae942a9284eecd41c8b28e09c3aed89ec363b50f4d2b420bf58b31181a766717dea8f814e74ad5f90d60fe4ca16985e67eee39b1e6ff2e9caee0ed9d4f33f8ac329f04806e547746ca8c61d5ea9f237b58d5d5240c07085bf710bb8d4ca6a61bf98a559155c3b1be81fb3853e292c53349b3e35b96f7d83f5f6781848d489d9cb5f4eb9874b342c4c625380c40e7771cb29b40ce9fff96b04e0f17822213253c9656bb4c7a7881284e56274c8b5c0de488cae76dbe6b4b06a2446d049aacf32d2fbf899bc2827e7080e899aca560f435db97724d0b428b0f9225df922f04f40c2b29d1a44310d72d26a558de437a652a3939afa4b79b388d0db5f9e5fa8849b98f628330b65068544731fce921a8dbb4e72200f31c75a9080fa9586dfce33f08d199bf6975dae325075fb663fc5ec8a9b4618dd9bac20fd1254a74e04bb1991376873644c9133d5a5ceadb8ce13f8a985f6951dc9b32c22fd44cf28e1d261e07fd32932b2d1067a6c2809d6922e5137d9dc79411c143d8b007ce4186b55aa62d43eb4c87441f4e6bbba9f7f7685e32f2dceddb767e5df80b3813a03eaaa5a41b89422dadbd868b6a6d45328d7d435e864864ce3177e4d1f1929913af3d5ae7305955ca1a0c2c83dd72a6b24da9f58335a99acf517b55d8ac60b5270abb8a7d0d248eb7fe444e0242ac2a16715483e827e529c3ee7badad9520bc762a16b85012863ee38f65434b75a97e8b5eed0962046947083606a7b4bbc2286e69a691aa38d2646c1075492a1ad105081b6518ef00560aca52c217b05fefbef8e461368310449f53017ea70a24ec5fbd88343b007dbd4df330eec040f65708fded16556811d6135ff5ffae94bc967fc98e8596411defe2bf5f67428f9494b0baa6149e5f13b31bbda2685877d2b4e903d36a10467edcd0e462bc6ca4255707d8f2a6fd881495138244d4de61577dfe9717f66c42d4001c00f1833bad8650c6394a42f7992b21abddfde21342218309f5152d7257fda8d8d9d2837e886b2ecd4baa62c7fc029420ad33bca144caffab917c545037e04dc8a28f32ba5994cb0f21bd6fd31300ca07248c0c152200afe43b7b6ef66086244793ac760add2869772840a90e91022e5431ee288b08c889f4b76a59aff8eea78bbbf2c600b674dda766439a01783837c9e9bba9ab9a0ac8a5aa226d7fc1d7f24116f81493a391b1f62332c20fe537c09f382a92e9a69ea77a5a1e7b83658797159a0bb0e546cedc8184cc1b79e002a929431ba4df75df16b7e55c62f996becc55cf544df4fb9f55b6aad67e70e93fbb23b73862ae41e680bc906c925dc6bebf9295b39d7e05d43e8bc797166782227cfb3243e6c1614197777b0d4b3333c986702d90f520989f4b9da0dd867cfa2dd559dbd0f602e183e247bb741d648f5aebeaaa23ce9c982af840e7ba4fa3053264165e615c7bbfed6a8aecbe9413890cb153d68819b553d5f9548a6b2b2b4e2d1239afc20bb7991114e55634e900a851c361133ad77bcc04a885d885147d331b54468cc0511fc0e1670f28fecbe5534043400483af7ce7a202076aa024ea762974c71f82ce9ff302abe57ce2d96f4d06f5540bae482ad83081aab35c78feaa3287f8b11df94bd4cabe80f938aef6ef44c5217d25015253240b216962ec5473101379d92c486deaf78e221b8af7cc27dfb2ac897aeceaf9a7434cc551ccc9b83c88dbbcaebbb4dc4c62af0626488c4dfb5612d35ef63d22f5266d2d919b767417e8712d450191c9b7bbfe9b157eef39e48a3671a703ab667f9074803cc2334012bd55997b11423a8b941c53113e645f98fecce3992015740650b4d212999ce6146d13cbdeca3d5e7c548c7e4f623a29e01af5cda44eb6edfa8706db63fff70c359e14bfd83412641d8b0abb829fa8ff9653a93e059170ad1b2eac9130c1498137f5be79ab12b331ead9e1f21972f1ce90e879108a17c4c08101f643d2bc33e71e85e37b4f1560144e3677bbaa1dd4ee49d351cf67a2bbe84703fcaae0da080f037a7af5946c1a891aebbfd690b502bd79cfd2829a31daa6660bcd936550d56d01fa2a46df83f7d3fdfd98d50d65c0c712ff274470d52cb4da1070cc4fa8bd8f3a110a87ba790727af6d09acad0c838cef1ccadf452e914feabb94474ec6c76d1085a276534ba835a490d923df662332debe0721e4fbc9ce1c3083d344c6268d1278b97c80c2dfb168ce332b86fc9a8141458cd0d704376dd8b1419c409291c229e5f8fc88f1668e7106b9cae31412610cc30e92a69a0ffd2743bdbf1746c22f18af6e7b4375d92b1ac20da26751f431f5803aff6a88a42f0e92cff9f254b0ee2bd16836010a0a12b4a14f99575fd589251fd6c1e2de3b6ee07483400c1ccd7337728301569069b1065a72bdef3bfb09835b9fd95e1caf1304c7162b4eabf27f196d246bf47f0ae46c07a51057bf5340483a5322244d6f03daba481250d7d6872de8d0934f98e5f77323ee571f9e0a75571be1c1a80cf037c42e7b848c8814913f65fc517c2b6e960f3e5a156f08bad12524966fd5e74b9d68128d85e0844009367733187962df33db1d196f2f64e823a585be79b81ded3afd4699c12c7a939e1e249e2747e624163c428001840c656b82389bae0a75f940d44c9aa03c54a576224cc659a4abac6a1a6b6e8ebd77ea02992043c28a80cd789826caf152a1f66818583c45dd32dbeb7066df37794c0c839500279916bdde7fae413d27df5141d529d3aaf2f3428929b8622a0a0bd46a6aed13883b30c453785f60d18f1bcdd10822d62083ab726a9fff5fdb2d83a2c89130ac91644985ae524d00509eac52c37818be1ac4c2bf9435de08a430957974cacdee6742ab172d2fa4b65871038301372dc58bfc9c9fd3db4aed3635d631518b20397695968047ed8672012760b122483cf9693207dbbaed34bb1e3bd498eb014b5055079fcc12b5f7e9eb46557f776623bbca8fc954b7eeb578af72b252b56fe307bb3de3720f7ba9b4f91d72ecc9fd27bf5d270f7945d12643d46a842b57d85498bf1717c58bbbdf65f84e50dd456e222eb2d6d7d4ee170e8f52e54157e87c28c8b317704183aea1b1fcf697b520a4283d9472d6cc4c3411853da19b048c15332e6f1b5f1e99684e927796d4418d3a3402f4fe1d1504f225881a59447b55c0a8197d28517895fe2c591e641a1576430ec4db6d928c6805138d8f6d7721e95402eb786dc491b4ef4542ed4f6074bb3c25fd092e1b74e6e295781b2a709a4a640cdfe965489e566f2f67f5f51aa45ca90a9b6c4f13faeda596ad5bfff69adf285296c13d8e43c421b4631eeb45c834980e2158f181cc572145b78889f1e85e0ea78434bb7b11bf7c8183c03ca8d5b3c71a641d5299a3f96292a11c29ac1bcc197cea126c5498eef000af2ddc6f82d2c41d284efcc30d43094a2689df863159b79e45b4918ec4678fb584da7985c9e4fa1b25d1b19cc926d5400f0d8aad919bfa755fbec2697474e8dc773b72d1f0177916b88830e758451bd0ee16c89ab8b42635cc54965148799f8424f0916b710a98001a61653255468ce5e43a5d471b3bf5f20f6f14a51d706810cc73c8bc1d3db33b6d4ad9e0d94387d129644635a2236a7475cde95baa78cd77b7d9b6281c53391ab1f1f03d459ffe12b427000eed1fc11533b3f7db80f19d38f426273595abeb316253e7b0db5ea4476fd868b11b3c764558376f5273738f950d313010c4b04c0569020452efa6956647de83358be0d9c27d449fb35fc354a88219bb4e77315c1ac10ff26110ae2057b9a54886ba40df3e84df041666e3c427a82f637abef62b1ae167808485367350b56eeb778d383470165b8b3e2a13b67db8cc5112f7e91132619ff5afbe56f5ebd5c7f86a73ccf04776683872664bc470b6288b0b162a87137c6b9f5b39726a97da51dd30001c20fd2ee5884506ada89485bb2f7c4a647c2e5da443ea22ece636039f1e82ca3105c8770f5ab98feeb1d31c11fb3e6cd59a2f75e1f7244a3939f82cd1682cc40f6a83d7d68da7a861dfe88ea12d2b0d76ed17b4211cd650789e785bac89a064eb85ce387575f71c7580a1e324fd276046ea0028637d005f9d633233045b58c8a586a99d29d38a18332b9fa7b7e3797cbd7a1a924bd75b3ddb7dcbf2593fc94b7726017f991b14bbc7ec0781c2b9f24e29d23013245e9a36c6b45adab1a2cf63956326d24c29b5db706ed93af785292c3f122a0cd55b3751c21eec9fd09a7ffaa00b06872e9333e97799c67de6efdad4499d1eec331973ff7d38d2c36f6e150ca547983bb21e8db3ef6f2173b98074bdf39e3d10f9e5d0cfe7f9d4d33cf2a16085bc5561c22b1f5a6c5902e65249d0ca94cf388fee83e6aafe1188a1402fa5f8ae8ba1840ea002be9ab3bd7b40fe235544d120770276f92c77883511bffb1a9a81ca5128c23c6fe6bf933f7364c2f50c2de7121f48536f779399bdc6fb928039d7b2287f8aff4c6edb5946af860479358e9f0813ee4cc117fbd838ae74f2989a5e6f64453bf18a7a421f58e3f76b461b2b63262020d564d298b115ea780b427f58de32eee1d96364bed3ca436f081a61a76625716cd2161b6a5992615c8eb30eb2a0497e0428c83a80d536b2258c06ff58253b1a2f82816857f414f270356fd728ea51966ace70dbe216a284f08f67b334733d5808c8d514decefbc68ff5d2f454a1045f27fdffa9bc743ae37ed2b7cfa8a83c77aeedcb47600c1dd9d062fb9afb11ba549a85fd75681fdbc8aa6817e7d8f7156e60ee016142063aca7ceb5501102b12dd4545804ee17f6697c03116c91c080bdc98bcc6ed122a19a59cc4f5bd266a0511982af91ab0ef1689557c90846fe71ff05cd8b5fa17eea64f9746d0590220dcb480619fd3b2fe7f6261c6cebb7ec5389913a82f61d410de5dda86467e02bb35a2a8f8ed7e796beaf21c8537e04c6e11f2be4adcf80d7f42bf80bc1782a1490a386cb9dc98a00f36d4409e3f0dde260d0181f59f53855a0ddd79ea94eb734e73dfbd640e78fa02f6e27258dac80ec22ea2cc30f52623fb23fd6183b4f84fd0a9641adb1dfac253e16c5da8d39faa6b308b455b85d7cb0b9b842fd5f975654b580809e98e20bb1071ce967d130852d42df5c3692b36a745306ed040c6f0a4dae0a0c4df54bf22fe6b6bfd94d400b4fc01fe8fa9b5fc060737c1b77585ec67a250084af4f9e87915b5af3e933dc1a62f1e32780535cd20e22e36a0535e0b2cd80af6784ece5d88bf77066b5ac7117154283bccdc82b704411a42690d6eb35f01ff005610165e8bb69333c46377ed730a01c8abd5a9cfc7ad4d537c78ee1270ba059675b9f2134034fefa1c9cd0453f31079975a7c727d1169f2d52e08f820d83f01eb5ec8a5c4b848e79b7759ba5a4c112da24374f938109a5207f076f3d463f8420b4ce29c7c2a1cb1012969e3901b20873839b1996074d26379ac62b40c15e5ca1dbb81dbb60133989f4b5ed20754d332c1b513454b89573f244abfe81d4cb7d30d203c7e5acf73e42001ebdd89f2b198d17e8fb48cd6e35666ed022130375bd4be9d446a7795ab9c5d99ac8704a0114b7b9de00aba48fd3cf97331709649d319d5f2c429f759a68cdeaa4a59d62c00e59c256325b85d674a964bb6a8a3a9dc5b4ed427d391a338b86b5f684984f0d28fbe7f0d4b1daf3e3927e75d0e55f4b00f88380e6e132ba6c9dedcefcb36f982d66b062c372b68a8127df8007389bbf537498f38c8d114ad2f72aad668452f233361a121d88c22247c853fc21f72d5e578cddfc06add0ef7b4cb0f7b841807723ce12d2fc7bae7eca2f0858bf7ec3b6010604e6f88bed5e5c0a6bb2ce701d0e347215d45f2b8b898bd36ef98c6c377496dd8aef60db4d647597f8a3aef16801324bc0905b9d0c539e50a8308a065557f273e5d0425d33e05be08e3f56a724052f4524979bdb63a384efb2e62fb56f9e3ebbb821605f0e00748f4b58fad456c10ddbe7f880977eb8f0ab9a901243eab679ffa6ee3069792cdd0cd2e06a5b34b71648fe85a919f1644acd3e8e2beac40b2a5483ce7807f4494427e0190ca5ad982c3292e961080305af60837411e86568790094edb46122ef5371ab4ecd11acf02c09ee4015d2f4fb80dfa0b309c653819ed5b599d0382eeb6a0caf77eb942b493cc182ee2884d4353794ee49587a7b6c3b9c093813422390e2ce9c5ceddd02452ee74eee4c9e67f70be1516d7167507aa649bbcf3726dffcfb596e36e24cbdc42e55d7b304e998822640af60d71c036c868fdcd6e99c19c35ff7b19fad62b4eb37447dfb62fb5a163ec13aeb3f1d4094487cfaf19f2f89647ee196904b50db8035b2d4fa59e533e3b7217af33b70b91d7fa8b00c1f67e4c3cd44addbb975865226b818b4dcec0b6d9d124d51d822101084cf8d52198c192a752a52d97ded183f30c9ed14866072551bdd39e9e33cb09204f72f728912767f835a9fac0c994b5631d74cbec8afe55311e8dadc62dc27c0a752b67df23526ace2fb4050e3a2c1163c4744940ebecbc0c04b965a255fc903d6610a82530a69b047ec935680f991f0186770e6713818ef573f38e9c57309f1bff77e973d49f096f372e55674f8a07901c564abb977c83be34baaf09762e472079f466eddf1b58659eedb1091c6c2e514fe91607d57afe0bedcd8bd6cf374f17f630496cfcbfe22122ca314398087bf3e214e734df59132914f92228bd5a463a57c682cdd6c9f8eedbac5f3f434e9d9218bc92f39a84bf769504f1f8947fafba65dcabd30696539f51e8407242849f50c660a7302951d618e09306401d9889b7c95b17980c1b953c0712728ee534c7308ae75492e6f861c4c65adfafeb157309228d46122b84d1bf3cc9857b15ee7630412aac1e07db4beef96aa55a33d63035bd3d5f8f6585a07763b17a0ebff2aa00c5ee96c960d6b86979fc8811d4e59e7756607971f3d289e2cd75be926d3c0842e202da7f7b5df43df989be9ac31c10364a9f26fad51a85c511389e6e4365d71129c23168a336d4c5b7cc674519fd8f91cbdacddf237352b726fd412cf75d5cb12eb789c284ab91c01394daae98fc4c60053c30641f9a906cd970f66ca7a27cf0838f09282c96e1e7a3368bd2b1eca98aebd797388e28daca5f9b59c4458c28a3be65f2022c45b6dd5d7efea2e442b672609b125ad2bb5775b02c73f63a8973f2a9e649cd10e2d38c616e51057392630315e6d767cf8e544b3dba9da47b8e69184f7817666e6aa539abe70b5ff086ef01c15ecd9b687d61cf9e2a07614c7233b1876cc8aadb5d24c210519b63e71a88fd0cac279883c3dd4a413533d231bbc701b1240b8554aaa633910599e68b27af0a5d189e222f34b91e7adf11db0e05e429f2e4bc0c80b21d218f23d2c7818f66272f3b94abd4ae71f0dfd7c14b6b1efcea3b10b9c5d915366942ae76b6052ad296ef41ea3995ee7544ca1d5b3331d461f33fa2b83d65f5ef8cb2e355bcaef6010cf567ab825d3aff17e906fafaa44c54d27717ae20d4d3f553125f085c91f52a3a9663881bd1abfc65c1aabede70a05d5326287473e2ceee2dafb5beeafbe3bbd8bf6aab2724f9c7916c9127676667bdeb2119962b10c7fd02247f8b565a4bfefe8d9bd80eb346aadda92e8c442f50ae6847f2d18b2209ba2142adddb1865f138c0667a1e24e0d17eb4d4576a2725fee0b2443dd21174f24a4114a6e5e4d930a2325889d57c1a98852dbd0596b0916b8e0ca68d4a998aa9675180a5bea087404432506b974c85b81fc6399f6ce1bb55660709eb10e99b889321a668bcb37ae47f53ef419ae75cb449439fc687a25e4c8c0673c0eb4465fa9cb1055fa3da4ecce22b2935709c79fecb6896e2d19489977c3479732e387e19e0a22a39979306b9d16df8750598373770093f05e856f6c878bfa5d13c89a2017c0539152620b87a37cf433f392e64bd50bddfdeb27cec0ece3e7bcbbdc0b8fd4c76cfe2eefa7055c7d01b85672b4c7280938369a4bc88cdc99c0e767a7daabd40a5c42b5a5f369622920dfbd8959afe059df38a2add2dc74c3efce90ff479455cfe359c5d67761a347b67d75b43b1ca3a9dea0e6e873a72f59e09fa0569550c264807959258dfde42783e4056494eb3ef19868c7e55b15d77b074ea94dfabd9f6c69899b806b09297e7feabde5d7eb655c6caef04eaff7296083c92ce8a5efc6a4002d4f8cedb31d1c5e393ad10344af91fce80b6840a6a3b3968b69eb35a1aff394de245981469c9687a15772260dbe5f4af27ecf85cbdeeae4bf2638e1bb2e6a92f3764bb75c79b590a85cc7fef175ead8538f7c04d23c52f5ed86a4baf5e80d965c5063907d01991aa90bbbeec72203f724f36d8e043a694a4510cc5dd3ad272df34ea90f0c1f29af70cd53ebaf4927d435a4c789213250ea6f7c8bf1b181d5ce565d496a72366dc0afac17b88bd862839f4ec6a8731f7820a9dad42589e626a2857fe98a1b38cbbbd46be5ffea8e3da9b09b6336d6fbe03",
-};
-
-export const [blockWithMyTx, blockWithoutMyTx, blockWithoutAnyTx] = blocks;
-export const [txNotShielded, txShieldedOrchard, txNotShielded2, txShieldedNotMine] = transactions;
