--- libs/coin-modules/zcash-shielded/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../../tsconfig.base",
-  "compilerOptions": {
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "declaration": true,
-    "declarationMap": true,
-    "lib": ["es2020", "dom"],
-    "types": ["node", "jest"],
-    "outDir": "lib",
-    "rootDir": "src",
-    "exactOptionalPropertyTypes": true
-  },
-  "include": ["src/**/*"]
-}
